### 2022-09-09 #4751 @editorialbot
[PRE REVIEW]: GNOLL: Efficient Multi-Lingual Software for Real-World Dice Notation and Extensions
**Submitting author:** @ianfhunter (<a href="http://orcid.org/0000-0003-3408-8138">Ian Hunter</a>)
**Repository:** https://github.com/ianfhunter/GNOLL
**Branch with paper.md** (empty if default branch):
**Version:** v2.7.1
