### 2022-09-12 edddc43
Create numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
