+++ doc/gh-pages/index.md
@@ -1,0 +1,10 @@
+Welcome to the GNOLL documentation!
+
+It is current under construction as we migrate from [our Wiki](https://github.com/ianfhunter/GNOLL/wiki)
+
+The aim of GNOLL is to provide the world with a library to make parsing dice notation as simple as possible, without compromising on functionality or performance. We too wish to provide direction on shaping dice notation standards.
+
+Want to set up GNOLL? Read our [Technical Setup](#)
+Want to understand dice notation? Start by understanding the basic types of dice and following [Grindon's Adventure](#).
+
+[Sponsor GNOLL](#)
