### 2022-09-12 479ae90
Update and rename _config.yml to doc/gh-pages/_config.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
