+++ doc/gh-pages/_config.yml
@@ -1,0 +1,5 @@
+theme: remote_theme: just-the-docs/just-the-docs
+exclude:
+ - src/*
+plugins:
+  - jekyll-seo-tag
