+++ doc/gh-pages/_config.yml
@@ -1,4 +1,5 @@
-theme: remote_theme: just-the-docs/just-the-docs
+# theme: 
+remote_theme: just-the-docs/just-the-docs
 exclude:
  - src/*
 plugins:
