### 2022-09-12 1f4c04b
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
