+++ doc/gh-pages/etc/further_reading.md
@@ -1,0 +1,9 @@
+---
+title: Further Reading
+published: true
+---
+
+A space for some interesting reads:
+
+- [The Origins of Dice Notation](http://playingattheworld.blogspot.com/2013/08/the-origins-of-dice-notation.html)
+- [Dice Patents](http://www.dicecollector.com/DICEINFO_PATENTS.html)
