+++ doc/gh-pages/notation/builtin_macros.md
@@ -1,0 +1,21 @@
+---
+title: Built-In Macros
+published: true
+---
+<!-- TODO: Grindon's Tale -->
+
+Here are a sample of the dice that can be used via Macro Accessors
+
+| Dice Type | Possible Values | Macro Accessor | 
+| --------- | --------------- | -------------- |
+| Poker Dice | 9, 10, J, K, Q, A | #POKER_DICE |
+| Chess Dice | Pawn, Knight, Bishop, Rook, Queen, and King | #CHESS_DICE |
+| Directions | NORTH, SOUTH, WEST, EAST | #DIRECTION_D4 |
+| Card Suits| CLUBS, DIAMONDS, HEARTS, SPADES | #CARD_SUIT |
+| Zodiac | Aries through Pisces | #ZODIAC |
+| Planets | Sun, Moon, Mercury through Pluto, Ascending Node, and Descending Node | #PLANETS |
+
+More can be found in the `builtins/` folder [here](https://github.com/ianfhunter/GNOLL/tree/main/builtins)
+
+To use a Macro, access it with '#' as below
+> #POKER_DICE
