### 2022-09-12 2bbb1b8
Update builtin_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
