+++ doc/gh-pages/notation/builtin_macros.md
@@ -6,6 +6,9 @@ published: true
 
 Here are a sample of the dice that can be used via Macro Accessors
 
+<!-- textlint-disable -->
+<!-- Ignore Suggestion for Node to be named Node.js -->
+
 | Dice Type | Possible Values | Macro Accessor | 
 | --------- | --------------- | -------------- |
 | Poker Dice | 9, 10, J, K, Q, A | #POKER_DICE |
@@ -14,6 +17,9 @@ Here are a sample of the dice that can be used via Macro Accessors
 | Card Suits| CLUBS, DIAMONDS, HEARTS, SPADES | #CARD_SUIT |
 | Zodiac | Aries through Pisces | #ZODIAC |
 | Planets | Sun, Moon, Mercury through Pluto, Ascending Node, and Descending Node | #PLANETS |
+
+
+<!-- textlint-enable -->
 
 More can be found in the `builtins/` folder [here](https://github.com/ianfhunter/GNOLL/tree/main/builtins)
 
