+++ doc/gh-pages/notation/builtin_macros.md
@@ -1,6 +1,8 @@
 ---
 title: Built-In Macros
 published: true
+
+nav_order: 9
 ---
 <!-- TODO: Grindon's Tale -->
 
