+++ doc/gh-pages/etc/further_reading.md
@@ -1,6 +1,7 @@
 ---
 title: Further Reading
 published: true
+nav_order: 11
 ---
 
 A space for some interesting reads:
