+++ doc/gh-pages/notation/drop_keep_notation.md
@@ -1,6 +1,8 @@
 ---
 title: Drop/Keep Notation
 published: true
+nav_order: 5
+
 ---
 
 # Grindon's Adventure
