### 2022-09-12 2912ca4
Update installation.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
