+++ doc/gh-pages/_config.yml
@@ -1,7 +1,10 @@
-
+# See: https://just-the-docs.github.io/just-the-docs/
 remote_theme: just-the-docs/just-the-docs
-exclude:
- - src/*
 plugins:
   - jekyll-seo-tag
   - jekyll-remote-theme
+
+# Search
+search_enabled: true
+search.heading_level: 3
+search.button: true
