### 2022-09-12 dd080b6
Update drop_keep_notation.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
