+++ tests/python/systems/dnd5E/test_5E.py
@@ -5,10 +5,14 @@ from util import roll, Mock
 
 def test_5e_roll_for_character_stats():
     # TODO: x6, for all stats
-    result = roll("4d6kh3")
+    result = roll("3d6")
     assert result >= 3
     assert result <= 18
 
+    result = roll("4d6kh3")
+    assert result >= 3
+    assert result <= 18
+    
 def test_5e_spell_attack():
     # TODO: Versus DC
     result = roll("1d20+2+4")
@@ -16,8 +20,10 @@ def test_5e_spell_attack():
     assert result <= 26
 
 def test_5e_roll_hp():
-    # For a level 2 character
     # TODO: ensure minimum roll of 1
+    
+    # Con:5 HitDie:8 Level:2
     result = roll("5+2+d8+2")
     assert result >= 10
     assert result <= 17
+
