### 2022-09-12 f240030
Create test_1E.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
