+++ doc/gh-pages/etc/further_reading.md
@@ -4,6 +4,8 @@ published: true
 nav_order: 11
 ---
 
+# Further Reading
+
 A space for some interesting reads:
 
 - [The Origins of Dice Notation](http://playingattheworld.blogspot.com/2013/08/the-origins-of-dice-notation.html)
