+++ doc/gh-pages/notation/symbolic_dice.md
@@ -5,8 +5,11 @@ nav_order: 2
 
 ---
 
-# Grindon's Tale 
+# Symbolic Dice
+
 ```
+   ⚔️ Grindon's Tale
+   
    Grindon: Hmm, I don't fancy getting in a brawl before I even enter the dungeon. And they might sound an alarm. I'll take a moment to look around for an alternate entrance.
    GM: Okay, Roll a Fate die and let's see
    Grindon the Brave: C'mon dice, work with me here! "d{-,0,+}"!
@@ -15,7 +18,6 @@ nav_order: 2
    GM: Sure enough, it seems like there's some sort of chimney hole about 20ft up from the main entrance. It looks big enough to squeeze through.
 ```
 
-# Symbolic Dice
 Symbolic dice have symbols on their sides instead of numbers. These could be letters, shapes or other iconography.
 
 To specify a symbolic dice, the symbols must be listed as there's no way for the parser to know what comes next. This syntax is very similar to custom numeric ranges. A simple example is d{A,B,C,D}.
