### 2022-09-13 19bd9d9
Update test_1E.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
