### 2022-09-13 07f93b1
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
