### 2022-09-13 5bf4f5d
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
