+++ doc/gh-pages/_config.yml
@@ -8,3 +8,17 @@ plugins:
 search_enabled: true
 search.heading_level: 3
 search.button: true
+
+# Callout blocks
+callouts:
+  warning:
+    title: ⚠️ Warning
+    color: red
+  
+  explanation:
+    title: 🤔 Design Decision
+    color: blue
+  
+  story:
+    title: ⚔️ Grindon's Tale
+    color: yellow
