+++ doc/gh-pages/notation/rerolling.md
@@ -6,17 +6,21 @@ nav_order: 7
 
 # Rerolling
 
-```
-   ⚔️ Grindon's Tale
-   
-   GM: Well, you took a load of damage, but you did level up.
-   Grindon: Nice, I get my class feature - I can reroll 1s!
-   GM: Do you remember how to do that?
-   Grindon: Yep, I just do a "1d20r#1"
-   GNOLL: [9]
-   GM: Great. Remember, you could still get a 1 on your second roll.
-   Grindon: Yeah, but then I deserved it.
-```
+{: .story}
+>
+>   GM: Well, you took a load of damage, but you did level up.
+>
+>   Grindon: Nice, I get my class feature - I can reroll 1s!
+>
+>   GM: Do you remember how to do that?
+>
+>   Grindon: Yep, I just do a "1d20r#1"
+>
+>   GNOLL: [9]
+>
+>   GM: Great. Remember, you could still get a 1 on your second roll.
+>
+>   Grindon: Yeah, but then I deserved it.
 
 Rerolling is often based on conditions (e.g. reroll if result is a 1). There are two options when using these conditions
  - Reroll once 
