+++ doc/gh-pages/notation/dice_shorthand.md
@@ -6,17 +6,21 @@ nav_order: 3
 
 # Dice Shorthand
 
-```
-   ⚔️ Grindon's Tale
 
-   Grindon: Alright, before I crawl through the chimney, I'm going to look down it to check it's safe
-   GM: There are the remains of a fire, but it looks safe to descend
-   Grindon: Great! Onwards!
-   GM: You crawl down the sooty chimney, getting well and truly covered in black dust. Stepping into the fireplace below, you find... d%
-   GNOLL: [74]
-   GM: *Consults Table* "A half-burnt wooden doll that looks eerily familiar. It looks like... you!"
-   Grindon: That's.. unsettling... I thought my ears were burning earlier.. 
-```
+{: .story }   
+> Grindon: Alright, before I crawl through the chimney, I'm going to look down it to check it's safe
+>
+> GM: There are the remains of a fire, but it looks safe to descend
+>
+> Grindon: Great! Onwards!
+>
+> GM: You crawl down the sooty chimney, getting well and truly covered in black dust. Stepping into the fireplace below, you find... d%
+>
+> GNOLL: [74]
+>
+> GM: *Consults Table* "A half-burnt wooden doll that looks eerily familiar. It looks like... you!"
+>
+> Grindon: That's.. unsettling... I thought my ears were burning earlier.. 
 
 ## Numeric Dice
 d% is often be used in place of a d100
