### 2022-09-13 2d88fd2
Update explosions.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
