+++ doc/gh-pages/_config.yml
@@ -22,3 +22,7 @@ callouts:
   story:
     title: ⚔️ Grindon's Tale
     color: yellow
+  
+  gnoll:
+    title: 🐺 GNOLL
+    color: brown
