### 2022-09-13 e4b69a0
Experiment with embedded callouts
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
