+++ doc/gh-pages/_config.yml
@@ -25,4 +25,4 @@ callouts:
   
   gnoll:
     title: 🐺 GNOLL
-    color: brown
+    color: grey-lt
