### 2022-09-13 a357010
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
