### 2022-09-13 00bad43
Create filters.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
