### 2022-09-13 c0149e7
Update dice_shorthand.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
