### 2022-09-13 007992a
Update test_5E.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
