### 2022-09-14 ec75da7
Update test_5E.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
