### 2022-09-17 5ff2d26
Update test_5E.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
