### 2022-09-17 af1e613
Create functions.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
