+++ doc/gh-pages/notation/functions.md
@@ -1,0 +1,11 @@
+---
+---
+
+#Functions
+
+Writing user functions is not implemented in GNOLL to avoid users who are not programming literate struggling with a foreign syntax.
+There are a limited set of predefined functions that can be used including:
+
+- clamp_high(<roll>, <clamp value>
+
+- clamp_low(<roll>, <clamp value>
