### 2022-09-17 62eeb62
Create test_DCC.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
