### 2022-09-17 1a78bc5
Update test_DCC.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
