### 2022-09-17 f5a0043
Update test_DCC.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
