+++ README.md
@@ -25,7 +25,7 @@ Here's an example of how you might use GNOLL:
    Dungeon Master: Hurrah! You successfully pickpocket the goblin! However, all he had in there were some crummy dice...
 ```
 
-[You can follow Grindon's full adventure through the world of dice notation in our Wiki](https://www.ianhunter.ie/GNOLL). 
+[You can follow Grindon's full adventure through the world of dice notation in our Documentation](https://www.ianhunter.ie/GNOLL). 
 
 Many of our notation design decisions are explained in the documentation and compared to other dice notation parsers.
 
