### 2022-09-20 123785a
Create test_troika.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
