### 2022-09-20 a63ac49
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
