### 2022-09-20 d6f7684
Create supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
