+++ doc/gh-pages/etc/supported_rpgs.md
@@ -1,0 +1,14 @@
+---
+title: Supported TTRPGs
+published: true
+nav_order: 11
+---
+# Supported TTRPGs
+
+(Under Construction)
+
+| TTRPG | Family of Games | Supported | Reference |
+| ----- | ---------- | --------- | ------- |
+| 13th Age | d20 | Yes | https://www.13thagesrd.com/combat-rules/ |
+
+* https://en.m.wikipedia.org/wiki/D20_System
