+++ doc/gh-pages/etc/supported_rpgs.md
@@ -10,5 +10,9 @@ nav_order: 11
 | TTRPG | Family of Games | Supported | Reference |
 | ----- | ---------- | --------- | ------- |
 | 13th Age | d20 | Yes | https://www.13thagesrd.com/combat-rules/ |
+| 2300AD | Custom | Likely | ? |
+| 3D&T, 4D&T | d20-like | Likely | ? |
+
+All items marked 'Likely' are unverified due to access constraints. However preliminary research has not raises any initial support gaps.
 
 * https://en.m.wikipedia.org/wiki/D20_System
