+++ doc/gh-pages/etc/supported_rpgs.md
@@ -12,6 +12,7 @@ nav_order: 11
 | 13th Age | d20 | Yes | https://www.13thagesrd.com/combat-rules/ |
 | 2300AD | Custom | Likely | ? |
 | 3D&T, 4D&T | d20-like | Likely | ? |
+| 7th Sea | Custom | Unknown | https://www.reddit.com/r/rpg/comments/47j3u8/7th_sea_system/ |
 
 All items marked 'Likely' are unverified due to access constraints. However preliminary research has not raises any initial support gaps.
 
