### 2022-09-21 bfa7f5e
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
