+++ doc/gh-pages/etc/supported_rpgs.md
@@ -13,6 +13,8 @@ nav_order: 11
 | 2300AD | Custom | Likely | ? |
 | 3D&T, 4D&T | d20-like | Likely | ? |
 | 7th Sea | Custom | Unknown | https://www.reddit.com/r/rpg/comments/47j3u8/7th_sea_system/ |
+| Aberrant | d20/Storyteller | Likely | https://en.m.wikipedia.org/wiki/Aberrant |
+| Active Exploits | Diceless | N/A | ? |
 
 All items marked 'Likely' are unverified due to access constraints. However preliminary research has not raises any initial support gaps.
 
