### 2022-09-21 a6a7758
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
