### 2022-09-21 8bad8fe
node support, split up makefiles, fix warnings
