+++ Makefile
@@ -1,19 +1,5 @@
-.PHONY: clean
-clean:
-	rm -rf build
-	rm src/grammar/test_dice.yacc | true
-# Python Packaging can create extraoneous files
-	rm src/python/src/ | true
-	rm src/python/dist/ | true
-	rm src/python/build/ | true
-	rm src/python/LICENSE | true
-	rm src/python/Makefile | true
-# Perl Artifacts
-	rm MYMETA.json | true
-	rm MYMETA.yml | true
-	rm result.die | true
 
-.PHONY: yacc
+.PHONY: clean yacc lex compile all test test_no_pip
 yacc:
 	mkdir -p build
 	if [ -z $(DEBUG) ]; then \
@@ -25,12 +11,10 @@ yacc:
 	mv y.tab.h build/y.tab.h
 	mv y.output build/y.output | true	# Only present with verbose
 
-.PHONY: lex
 lex:
 	lex src/grammar/dice.lex
 	mv lex.yy.c build/lex.yy.c
 
-.PHONY: compile
 compile:
 # Executable
 	cc -O3 build/y.tab.c \
@@ -57,104 +41,21 @@ compile:
 # Windows
 	mv ./a.exe build/dice | true
 
-.PHONY: all
 all: clean yacc lex compile
 	echo "== Build Complete =="
 
-.PHONY: test_no_pip
 test_no_pip :
 	python3 -m pytest tests/python/ -x
 
-.PHONY: test
 test : pip
 	python3 -m pytest tests/python/ -x
 
-.PHONY: pip
-pip : all
-	echo "----------------- BUILD -------------------------"
-	# Copy Build
-	rm -rf src/python/code/gnoll/c_build/
-	rm -rf src/python/code/gnoll/c_includes/
-	rm -rf src/python/code/gnoll.egg-info/
+include src/python/target.mk
+include src/js/target.mk
+include src/go/target.mk
+include src/perl/target.mk
+include src/swig/target.mk
 
-	cp -r build/  src/python/code/gnoll/c_build/
-	cp -r src/grammar/  src/python/code/gnoll/c_includes/
-
-	cd src/python/ ; python3 -m build
-	echo "------------------INSTALL------------------------"
-	python3 -m pip install -vvv --user --no-index --find-links=src/python/dist/ --force-reinstall --ignore-installed gnoll
-	echo "-------------------- TEST ----------------------"
-	python3 -c "from gnoll import parser as dt; dt.roll('2')"
-
-
-.PHONY: publish
-publish: test
-	#twine upload --repository-url https://test.pypi.org/legacy/ src/python/dist/*
-	twine upload src/python/dist/*
-
-# An example
-.PHONY: swig
-swig: swig_perl swig_java swig_go swig_js
-
-.PHONY: swig_go
-swig_go:
-	mkdir -p build/go/
-	swig -go -outdir build/go -o src/go/gnoll_wrap.c src/swig/GNOLL.i
-
-.PHONY: swig_java
-swig_java:
-	mkdir -p build/java/
-	swig -java -outdir build/java -o src/java/gnoll_wrap.c src/swig/GNOLL.i
-
-.PHONY: swig_js
-swig_js:
-	mkdir -p build/js/
-	swig -javascript -outdir build/js -o src/js/gnoll_wrap.c src/swig/GNOLL.i
-
-
-# ========= PHP =========
-# Note: SWIG only supports PHP 5.0 - 5.3 which is very outdated
-# Current version at time of writing is 8.0
-
-# ========= PERL =========
[…]
