### 2022-09-21 63b36ef
fix: emscripten not on 20.04+ ubunuts
