+++ .github/workflows/js-test.yml
@@ -14,6 +14,9 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: javascript prerequisites
-      run: sudo apt install nodejs emscripten
+      run: |
+        sudo apt install nodejs 
+        wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten_1.22.1-1build_amd64.deb
+        sudo apt install ./emscripten_1.22.1-1build_amd64.deb
     - name: make js
-      run: make ks
+      run: make js
