### 2022-09-21 b96d85f
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
