+++ .github/workflows/js-test.yml
@@ -16,7 +16,7 @@ jobs:
     - name: javascript prerequisites
       run: |
         sudo apt install nodejs 
-        wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten_1.22.1-1build_amd64.deb
-        sudo apt install ./emscripten_1.22.1-1build_amd64.deb
+        wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten_1.22.1-1build1_amd64.deb
+        sudo apt install ./emscripten_1.22.1-1build1_amd64.deb
     - name: make js
       run: make js
