+++ doc/gh-pages/etc/supported_rpgs.md
@@ -7,42 +7,49 @@ nav_order: 11
 
 (Under Construction) List from https://en.m.wikipedia.org/wiki/List_of_tabletop_role-playing_games
 
-| TTRPG | Family of Games | Supported | Reference |
-| ----- | ---------- | --------- | ------- |
-| 13th Age | d20 | Yes | https://www.13thagesrd.com/combat-rules/ |
-| 2300AD | Custom | Likely | ? |
-| 3D&T, 4D&T | d20-like | Likely | ? |
-| 7th Sea | Custom | Unknown | https://www.reddit.com/r/rpg/comments/47j3u8/7th_sea_system/ |
-| Aberrant | d20/Storyteller | Likely | https://en.m.wikipedia.org/wiki/Aberrant |
-| Active Exploits | Diceless | N/A | ? |
-| Advanced Dungeons & Dragons | D&D | Yes | ? |
-| Dungeons & Dragons 3.5 | D&D | Yes | ? |
-| Dungeons & Dragons 4 | D&D | Yes | ? |
-| Dungeons & Dragons 5 | D&D | Yes | ? |
-| Advanced Fighting Fantasy | - | Yes | ? |
-| Adventure! (1985) | ? | Unknown | None Found |
-| Adventure! (2001) | Modified Storyteller | Likely | ? |
-| Adventurer's Guild | ? | Unknown | None Found |
-| Adventures in Fantasy | ? | Likely | https://www.scribd.com/document/245240865/1979-Adventures-in-Fantasy-Book-of-Adventure |
-| The Adventures of Indiana Jones | ? | Likely | |
-| Adventures in Middle Earth | d20 | Yes | |
-| After the bomb | Megaversal | Yes | |
-| Aftermath | Custom | Likely | https://forum.rpg.net/index.php?threads/aftermath-where-i-read-the-aftermath-rules.684493/ |
-| Age of the Tempest | Custom | Unknown | None Found |
-| Age of Ruin | Custom | Likely | None Found |
-| Agone | d10 | Likely | |
-| Albeldo | Custom | Unknown | None Found |
-| Alien | Custom | Likely | |
-| Aliens | Custom | Likely | |
-| All flesh must be eaten | Unisystem | Likely | |
-| All-Star Wrestling | Custom | Likely | |
-| Alma Mater | Custom | Unknown | |
-| Alshard | Custom | Yes ||
-| Alshard Gaia | Custom | Yes ||
-| Alternity | Custom | Likely ||
+## Legend
 
+| Symbol | Meaning |
+| ------ | ------- |
+| ✔️     | Supported |
+| 🔼     | Likely Supported (Unable to fully verify because of lack of access to rules, etc) |
+| ❓      | Unknown (Unable to find details on ruleset |
+| 🔽     | Unlikely Supported (Unable to fully verify because of lack of access to rules, etc) |
+| ❌     | Unsupported |
+| 〰️     | Not Applicable |
 
+## List
 
-All items marked 'Likely' are unverified due to access constraints. However preliminary research has not raises any initial support gaps.
-
-* https://en.m.wikipedia.org/wiki/D20_System
+| TTRPG | Family of Games | Supported | 
+| ----- | ---------- | --------- | 
+| 13th Age [[1]](https://www.13thagesrd.com/combat-rules/) | d20 | ✔️ | 
+| 2300AD |  | 🔼 |
+| 3D&T, 4D&T | d20-like | 🔼 |
[…]
