+++ tests/python/systems/DCC/test_DCC.py
@@ -28,7 +28,7 @@ def test_dcc_weird_dice():
 
 def test_dcc_weird_dice_via_other_dice():
     # https://www.reddit.com/r/rpg/comments/79lge5/weird_dice_in_dcc/
-    result = roll("d10/2")
+    result = roll(r"d10\2")
     assert result >= 1
     assert result <= 5
     result = roll("d6+(6*z2)")
