### 2022-09-22 e888f2a
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
