+++ .github/workflows/c-cpp.yml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   build:
-
+    name: "Test: C/C++/Python"
     runs-on: ubuntu-latest
 
     steps:
