### 2022-09-22 6117349
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
