+++ doc/gh-pages/etc/supported_rpgs.md
@@ -92,6 +92,8 @@ BCDEFGHIJKLMNOPRSTVW
 | Atlantis |  | ❓ |
 | Attack of the Humans |  | ❓ |
 | The Authority Role-Playing Game |  | 🔼 |	
+| Year of the Phoenix | | 🔼 |
+| Ysgarth| | 🔼 |
 | The Zorcerer of Zo | PDQ | ✔️ |
 | Zweihänder [[1]](https://1d4chan.org/wiki/Zweihander) | | ✔️ |
 
