### 2022-09-23 ee5dcd0
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
