### 2022-09-23 967fbf0
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
