### 2022-09-23 1e4fee8
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
