### 2022-09-23 4ce85a7
Create GNOLL.ini
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
