+++ GNOLL.ini
@@ -1,0 +1,4 @@
+; Placeholder file for future configuration setting. Currently unused
+[Meta Information]
+version = v2.8.0
+ttrpg_compatibility = 96.875
