### 2022-09-23 2a5eaed
Update GNOLL.ini
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
