### 2022-09-23 b52ac92
JS Test should roll dice not just build
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
