### 2022-09-24 ad78c27
Update setup.cfg
Signed-off-by: jay-jpj <79367341+jay-jpj@users.noreply.github.com>
