+++ src/python/setup.cfg
@@ -18,7 +18,7 @@ classifiers =
     Operating System :: Unix
     Intended Audience :: Developers
     Intended Audience :: Science/Research
-    Intended Audience :: Other Audience
+    Intended Audience :: Any other Audience
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
     Topic :: Games/Entertainment :: Board Games
     Topic :: Games/Entertainment :: Role-Playing
