### 2022-09-24 73533fd
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
