+++ .github/workflows/js-test.yml
@@ -20,7 +20,6 @@ jobs:
         sudo apt install ./emscripten_1.22.1-1build1_amd64.deb
     - name: make js
       run: |
-        emscripten -v
         make js
         ./src/js/a.out.js 1d20
         
