### 2022-09-24 5040cb1
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
