### 2022-09-24 e7e99ea
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
