### 2022-09-24 ec66123
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
