### 2022-09-24 dad2b5b
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
