### 2022-09-24 e38485e
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
