### 2022-09-24 dda3b92
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
