### 2022-09-24 be294be
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
