### 2022-09-24 dccfb57
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
