### 2022-09-24 04b3ae4
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
