### 2022-09-24 b0492d0
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
