### 2022-09-25 964c561
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
