### 2022-09-25 401b971
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
