### 2022-09-25 60c57c5
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
