### 2022-09-26 ad92d28
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
