### 2022-09-26 99ac201
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
