### 2022-09-26 d386c58
Improvements to success counting documentation
