+++ doc/gh-pages/notation/success_counting.md
@@ -1,19 +1,0 @@
----
-title: Success Counting
-published: true
-nav_order: 8
----
-
-# Success Counting
-
-{: .story}
-> TODO
-
-*Under Construction*
-
-## Counting Successes
-
-In several games, the rolls may not be a matter of *if* your total of rolls is good, but *how many* good rolls you have.
-For this, we count the number of successes, according to a criteria. By default a max-face value is assumed to be a success, but you can change this, much like other condition changes.
-
-> 100dFc - Count how many Fate dice are +
