### 2022-09-26 e685cfa
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
