+++ src/python/setup.cfg
@@ -34,6 +34,8 @@ package_dir =
 packages = find_namespace:
 python_requires = >=3.6
 include_package_data=True
+install_requires=
+    cppyy
 
 [options.packages.find]
 where = code
