### 2022-09-26 d859a7d
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
