### 2022-09-26 2f3e4ae
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
