### 2022-09-27 a47b179
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
