+++ src/python/setup.cfg
@@ -35,7 +35,7 @@ packages = find_namespace:
 python_requires = >=3.6
 include_package_data=True
 install_requires=
-    cppyy
+    cppyy>=2.0.0
 
 [options.packages.find]
 where = code
