### 2022-09-27 3fb91e9
Create dice_pools.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
