+++ doc/gh-pages/notation/dice_pools.md
@@ -1,0 +1,17 @@
+---
+title: Dice Pools
+published: true
+nav_order: 3
+---
+
+# Dice Pools
+
+```
+{: .story }
+Todo
+```
+
+## Examples
+
+ - `1,2` 
+ - `d20, d20`
