### 2022-09-27 72bc233
Update pyproject.toml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
