### 2022-09-27 9c50da1
Update pyproject.toml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
