### 2022-09-27 8c53d41
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
