+++ src/python/setup.cfg
@@ -34,8 +34,8 @@ package_dir =
 packages = find_namespace:
 python_requires = >=3.6
 include_package_data=True
-install_requires=
-    cppyy>=2.0.0
+# install_requires=
+#    cppyy>=2.0.0
 
 [options.packages.find]
 where = code
