### 2022-09-27 10bec59
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
