### 2022-09-27 8e9747d
Update pyproject.toml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
