+++ .github/pull_request_template.md
@@ -20,6 +20,8 @@
 - [ ] Bug fix (non-breaking change which fixes an issue)
 - [ ] New feature (non-breaking change which adds functionality)
 - [ ] Breaking change (fix or feature that would cause existing functionality to change)
+- [ ] Test Coverage (Additional test(s) without any extra code)
+- [ ] Documentation Improvements 
 
 ## Checklist:
 <!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
