### 2022-09-27 b41d401
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
