+++ doc/gh-pages/etc/supported_rpgs.md
@@ -160,6 +160,7 @@ This list is currently **Under Construction**.
 |d20 System	| D20 | ✔️ |	
 | Dallas [[1]](https://writeups.letsyouandhimfight.com/bieeardo/dallas-rpg/) | 	| ✔️ |
 | Danger International | Hero | ✔️ |
+| Dangerous Journeys | | ✔️ |
 | Unisystem | Unisystem | ✔️ |
 | Universalis | | ✔️|
 | Universe | | ✔️ [#203](https://github.com/ianfhunter/GNOLL/issues/203)
