### 2022-09-27 841c605
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
