+++ .github/workflows/c-cpp.yml
@@ -46,7 +46,11 @@ jobs:
     - uses: actions/checkout@v2
     - name: python deps
       shell: wsl-bash {0}
-      run: python3 -m pip install -r reqs/requirements.txt
-    - name: make test
+      run: |
+        python3 -m pip install --upgrade pip
+        python3 -m pip install -r reqs/requirements.txt
+        python3 -m pip install build
+        make all
+    - name: test
       shell: wsl-bash {0}
       run: make test
