### 2022-09-27 50311c9
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
