+++ .github/workflows/c-cpp.yml
@@ -42,6 +42,9 @@ jobs:
           bison
           flex
           python3-pip
+          build-essential 
+          libssl-dev 
+          libffi-dev
     
     - uses: actions/checkout@v2
     - name: python deps
