### 2022-09-27 07439f3
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
