+++ src/python/setup.cfg
@@ -2,7 +2,7 @@
 name = gnoll
 author = Ian Frederick Vigogne Goodbody Hunter
 author_email = ianfhunter@gmail.com
-version = 2.8.3
+version = 2.8.4
 description = An efficient dice notation parser with extended notation
 long_description = file: README.md
 long_description_content_type = text/markdown
