### 2022-09-28 82e91ac
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
