+++ .github/workflows/c-cpp.yml
@@ -23,9 +23,7 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip
         python -m pip install -r reqs/requirements.txt
-        pip install build
         make all
     - name: make test
       run: make test
