### 2022-09-28 a0d023d
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
