### 2022-09-28 3e45937
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
