### 2022-09-28 445624d
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
