+++ .github/workflows/c-cpp.yml
@@ -35,6 +35,7 @@ jobs:
     steps:
     - uses: Vampire/setup-wsl@v1
       with:
+        distribution: Ubuntu-22.04
         additional-packages:
           python3-dev
           bison
