### 2022-09-28 63b3865
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
