+++ .github/workflows/c-cpp.yml
@@ -49,12 +49,12 @@ jobs:
     - name: python deps
       shell: wsl-bash {0}
       run: |
-        export DISPLAY=
         unset DISPLAY 
-        python3 -m pip install --upgrade pip
-        python3 -m pip install -r reqs/requirements.txt
-        python3 -m pip install build
-        make all
+        DISPLAY=  python3 -m pip install -r reqs/requirements.txt
+        
+    - name: build
+      shell: wsl-bash {0}
+      run: make all
     - name: test
       shell: wsl-bash {0}
       run: make test
