### 2022-09-28 7839f74
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
