### 2022-09-28 179ba40
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
