### 2022-09-28 f8df9e6
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
