### 2022-09-28 25a078b
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
