### 2022-09-28 107fc50
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
