### 2022-09-28 bdec60a
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
