### 2022-09-28 c32b936
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
