+++ .github/workflows/c-cpp.yml
@@ -53,6 +53,7 @@ jobs:
         unset DISPLAY 
         python3 -m venv env
         source env/bin/activate
+        python3 -m pip --upgrade pip
         python3 -m pip -v install -r reqs/requirements.txt
         
     - name: test
