### 2022-09-28 cf271fb
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
