### 2022-09-28 a531e8e
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
