### 2022-09-28 ca20ef7
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
