### 2022-09-28 036d170
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
