+++ .github/workflows/c-cpp.yml
@@ -54,9 +54,9 @@ jobs:
         python3 -m venv env
         source env/bin/activate
         python3 -m pip install --upgrade pip
-        # python3 -m pip -vvv --no-cache-dir install -r reqs/requirements.txt
-        python3 -m pip install cppyy
-        python3 -m pip install pytest
+        # python3 -m pip install -r reqs/requirements.txt
+        python3 -m pip  -vvv --no-cache-dir  install cppyy
+        python3 -m pip  -vvv --no-cache-dir  install pytest
         
     - name: test
       shell: wsl-bash {0}
