### 2022-09-28 d3a3c6e
Update c-cpp.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
