+++ .github/workflows/c-cpp.yml
@@ -53,7 +53,7 @@ jobs:
         unset DISPLAY 
         python3 -m venv env
         source env/bin/activate
-        python3 -m pip install --upgrade pip
+        python3 -m pip  -vvv --no-cache-dir install --upgrade pip
         # python3 -m pip install -r reqs/requirements.txt
         python3 -m pip  -vvv --no-cache-dir  install cppyy
         python3 -m pip  -vvv --no-cache-dir  install pytest
