+++ .github/workflows/js-test.yml
@@ -22,8 +22,8 @@ jobs:
     - name: javascript prerequisites
       run: |
         sudo apt install nodejs llvm -y
-        wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten_1.22.1-1build1_amd64.deb
-        sudo apt install ./emscripten_1.22.1-1build1_amd64.deb
+        wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten-doc_3.1.6~dfsg-5_all.deb
+        sudo apt install ./emscripten-doc_3.1.6~dfsg-5_all.deb
     - name: make js
       run: |
         make all
