### 2022-10-01 abecc25
Changed the character name to Bold and  GNOLL to Italic
