+++ doc/gh-pages/notation/drop_keep_notation.md
@@ -9,17 +9,17 @@ nav_order: 5
 
 
 {: .story }   
->   GM: Alright, you step over the goblin's corpse and go inside. There's another one there. He hasn't seen you yet.
+>   **GM**: Alright, you step over the goblin's corpse and go inside. There's another one there. He hasn't seen you yet.
 >
->Grindon The Brave: I guess I should have talked to the other one, you were right. This one i will just stealth and put him to sleep.
+>**Grindon The Brave**: I guess I should have talked to the other one, you were right. This one i will just stealth and put him to sleep.
 >
->   GM: It's pretty dark in here, so I'll give you advantage on a stealth roll
+>   **GM**: It's pretty dark in here, so I'll give you advantage on a stealth roll
 >
->   Grindon The Brave: Nice, "2d20kh"
+>   **Grindon The Brave**: Nice, "2d20kh"
 >
->   GNOLL: [20]
+>   *GNOLL*: [20]
 >
->   GM: You are the night itself. You take down the goblin before it can even react. 
+>   **GM**: You are the night itself. You take down the goblin before it can even react. 
 ```
 
 ## Implementation Comparison
