### 2022-10-01 2f16954
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
