### 2022-10-01 dadc568
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
