### 2022-10-01 c3dd204
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
