+++ .github/workflows/python-publish.yml
@@ -27,7 +27,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install build
-        make all
+        make python
         ls src/python
         echo "-------"
         ls src/python/code/gnoll/
