### 2022-10-01 7411c20
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
