### 2022-10-01 01f3601
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
