### 2022-10-01 373b3e0
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
