### 2022-10-01 236e55d
Create operations.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
