### 2022-10-02 6f1b239
Update operations.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
