### 2022-10-02 b0dbf4a
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
