### 2022-10-02 004c2ef
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
