### 2022-10-02 7319fd7
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
