### 2022-10-02 6e5b32f
Update example_application.pl
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
