+++ src/perl/example_application.pl
@@ -4,9 +4,6 @@ use gnoll;
 
 print "GNOLL in PERL\n";
 
-# my $result = gnoll::roll_verbose("1d100+3");
-# print $result;
-
 my $result_file = 'result.die';
 unlink($result_file);
 my $result = gnoll::roll_and_write("1d100+3", $result_file);
@@ -18,4 +15,3 @@ while(<FH>){
 close(FH);
 
 print "\n";
-;
