### 2022-10-02 db8bc6d
Update example_application.pl
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
