+++ src/perl/example_application.pl
@@ -1,6 +1,6 @@
 use File::Basename;
-use lib dirname (__FILE__);
 use gnoll;
+use lib dirname (__FILE__);
 
 print "GNOLL in PERL\n";
 
@@ -8,10 +8,10 @@ my $result_file = 'result.die';
 unlink($result_file);
 my $result = gnoll::roll_and_write("1d100+3", $result_file);
 
-open(FH, '<', $result_file) or die $!;
-while(<FH>){
+open(my $FH, '<', $result_file) or die $!;
+while(<$FH>){
    print $_;
 }
-close(FH);
+close($FH);
 
 print "\n";
