### 2022-10-02 1a79565
Update example_application.pl
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
