### 2022-10-03 #4816 @galessiorob
👋 @Smattr, please create your checklist typing: `@editorialbot generate my checklist`

This will give you an itemized list for you to follow. Please let me know if you have any questions.
