### 2022-10-03 9aee4e1
Update dice_counting.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
