+++ doc/gh-pages/notation/dice_counting.md
@@ -25,14 +25,22 @@ nav_order: 8
 >
 >   Grindon: Yeah, but then I deserved it.
 
-## Counting Successes
+## Counting Dice
 
 In several games, the rolls may not be a matter of *if* your total of rolls is good, but *how many* good rolls you have.
 For this, we count the number of successes, according to a criteria. By default a max-face value is assumed to be a success, but you can change this, much like other condition changes.
 
-> 10d10f>7 - Count how many D10s are larger than 7
+> 10d10f>7c - Count how many D10s are larger than 7
 
 > 100dFc - Count how many Fate dice are +
+
+{: .explaination }
+> Counting dice is a rarer feature of dice rollers and is sometimes used with the # symbol [[1]](https://www.sophiehoulden.com/dice/documentation/notation.html#count)
+> Others merge the functionality with filtering[[2]](https://www.critdice.com/blog/2016/10/30/critdice-version-20-released).
+>
+> As we already use # for Macro Setting, we chose to use 'c' standing for 'count' which we believe is more straightforward anyhow.
+
+## Counting Successes
 
 {: .explanation }   
 > Some games have a specific concept of success counting
