### 2022-10-03 9b1a475
Update builtin_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
