### 2022-10-03 39ecf9c
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
