### 2022-10-04 541d3e9
Update installation.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
