### 2022-10-04 6699292
Update .markdown-lint.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
