+++ .github/linters/.markdown-lint.yml
@@ -33,3 +33,9 @@ MD036: false                  # Emphasis used instead of a heading
 # Rules by tags #
 #################
 blank_lines: false  # Error on blank lines
+
+
+############################
+# GNOLL Rule Modifications # 
+############################
+MD025: false                  # Incompatible rule with Jekyll Setup
