### 2022-10-04 9eff15b
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
