+++ doc/gh-pages/index.md
@@ -1,14 +1,11 @@
 Welcome to GNOLL's documentation!
 
-Please bear with us as we migrate from [our Wiki](https://github.com/ianfhunter/GNOLL/wiki).
-
 The aim of GNOLL is to provide the world with a dice parsing library that:
 - Is simple to use (By the general public, not by statisticians)
 - Is easy to integrate into up-and-coming software projects
 - Is performant
 - Supports a wide range of dice notation formats and operations.
 - Can be used as a reference implementation for others trying to understand dice notation.
-
 
 If you're looking to set up GNOLL, read our [Technical Setup](setup/installation.md)
 
