### 2022-10-04 0e12088
Update explosions.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
