+++ doc/gh-pages/notation/explosions.md
@@ -5,26 +5,27 @@ nav_order: 6
 ---
 
 # Explosions & Implosions
-{: .story }   
->  **Grindon**: Oh gosh, uh. What's in the room?! How close is the gnoll?
+
+{: .story } 
+>**Grindon**: Oh gosh, uh. What's in the room?! How close is the gnoll?
 >
->   **GM**: He's just coming through the door. There's a small table, a chest and some barrels of oil.
+>**GM**: He's just coming through the door. There's a small table, a chest and some barrels of oil.
 >
->   **Grindon**: Oil! I light a torch and throw a match, diving back into the fireplace.
+>**Grindon**: Oil! I light a torch and throw a match, diving back into the fireplace.
 >
->   **GM**: This is gonna be a close one! Roll an athletics check to get into the fireplace and an exploding d6 for the barrel of oil.
+>**GM**: This is gonna be a close one! Roll an athletics check to get into the fireplace and an exploding d6 for the barrel of oil.
 >
->   **Grindon**: Okay that's... 1d20+4 then... d6!
+>**Grindon**: Okay that's... 1d20+4 then... d6!
 >
->   *GNOLL*: [18]
+>*GNOLL*: [18]
 >
->   *GNOLL*: [25]
+>*GNOLL*: [25]
 >
->   **GM**: The explosion blasts behind you as you dive inside. You still take 4 damage as you smash into the stone and your ears ring as the barrel bursts, evaporating the gnoll.
+>**GM**: The explosion blasts behind you as you dive inside. You still take 4 damage as you smash into the stone and your ears ring as the barrel bursts, evaporating the gnoll.
 >
->   **Grindon The Brave**: Ouch! 
+>**Grindon The Brave**: Ouch! 
 >
->   **GM**: I'm pretty sure whatever else is in this cave heard that too :)
+>**GM**: I'm pretty sure whatever else is in this cave heard that too :)
 
 
 ## Traditional Explosions
@@ -55,44 +56,29 @@ When an explosion is set off, the range of the next dice is reduced by 1 step. A
 
 Not Implemented Yet. See [Issue #11](https://github.com/ianfhunter/GNOLL/issues/11) 
 
+Due to the more complicated notation, and the likelyhood that a penetrating explosion is standard across an rpg system, we recommend saving these rolls as macros.
+
 
 ## Diminishing Explosions
 
-*Under Construction*
+Diminishing Explosions can be thought of the reverse of Penetrating Explosions. When an explosion is set off, we roll a smaller dice each time.
 
-<!---
-When an explosion is set off, we roll a smaller dice each time.
-e.g. 1d20,1d12,d10,d8,d6,d4
-
-As this system is not locked into particular rpg sets, you must define these in a variable. This also gives the option of incrementing explosions and non-directional explosions.
-> d20![d12,d10,d8,d6,d4]
-
-We recommend creating a macro for common use cases:
-$DND_5E_EXPLOSION_SET=d12,d10,d8,d6,d4 ; d20![$DND_5E_EXPLOSION_SET]
-
-> What the dice here was roll several max values (3,3,3,3) and finally a 1. A max value triggers another roll to be added on.
--->
+GNOLL will implement this through the same syntax as Penetrating Explosions.
 
 ## Imploding dice
 
 *Under Construction*
 
-<!--
 There is no consistent standard for imploding dice. Here are a few candidates:
- - [Inverse Explosion - Rerolling when not matching your number](https://www.welshpiper.com/imploding-dice/)
- - [Negative Explosion](http://www.firehawkgames.biz/?qa_faqs=what-is-an-imploding-dice-roll)
--->
[…]
