### 2022-10-04 ef5faae
Update builtin_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
