+++ doc/gh-pages/notation/builtin_macros.md
@@ -7,7 +7,6 @@ nav_order: 9
 
 # Built-In Macros
 
-
 {: .story }   
 > TODO
 
@@ -21,7 +20,7 @@ Here are a sample of the dice that can be used via Macro Accessors
 | Directions | NORTH, SOUTH, WEST, EAST | #DIRECTION_D4 |
 | Card Suits| CLUBS, DIAMONDS, HEARTS, SPADES | #CARD_SUIT |
 | Zodiac | Aries through Pisces | #ZODIAC |
-| Planets | Sun, Moon, Mercury through Pluto, Ascending Node, and Descending Node | #PLANETS |
+| Planets | Sun, Moon, Mercury through Pluto, Ascending Node, Descending Node | #PLANETS |
 
 More can be found in the `builtins/` folder [here](https://github.com/ianfhunter/GNOLL/tree/main/builtins)
 
