### 2022-10-04 7c5e610
Update custom_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
