+++ doc/gh-pages/notation/custom_macros.md
@@ -12,6 +12,7 @@ nav_order: 10
 
 
 ## Usage
+
 *This is an advanced feature. Feel free to ask us for help!*
 
 There are certain system-specific [[Built-In Macros]] for you to take advantage of.
