### 2022-10-04 6538419
Update dice_counting.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
