### 2022-10-04 e19deea
Update dice_shorthand.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
