### 2022-10-04 d898c7d
Update drop_keep_notation.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
