### 2022-10-04 9f55f95
Update symbolic_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
