+++ doc/gh-pages/notation/symbolic_dice.md
@@ -9,17 +9,17 @@ nav_order: 2
 
 {: .story}
 >
->   **Grindon**: Hmm, I don't fancy getting in a brawl before I even enter the dungeon. And they might sound an alarm. I'll take a moment to look around for an alternate entrance.
+>**Grindon**: Hmm, I don't fancy getting in a brawl before I even enter the dungeon. And they might sound an alarm. I'll take a moment to look around for an alternate entrance.
 >
->   **GM**: Okay, Roll a Fate die and let's see
+>**GM**: Okay, Roll a Fate die and let's see
 >
->   **Grindon** the Brave: C'mon dice, work with me here! "d{-,0,+}"!
+>**Grindon** the Brave: C'mon dice, work with me here! "d{-,0,+}"!
 >
->   *GNOLL*: ["+"]
+>*GNOLL*: ["+"]
 >
->   **Grindon the Brave**: Yes! 
+>**Grindon the Brave**: Yes! 
 >
->   **GM**: Sure enough, it seems like there's some sort of chimney hole about 20ft up from the main entrance. It looks big enough to squeeze through.
+>**GM**: Sure enough, it seems like there's some sort of chimney hole about 20ft up from the main entrance. It looks big enough to squeeze through.
 
 
 Symbolic dice have symbols on their sides instead of numbers. These could be letters, shapes or other iconography.
