### 2022-10-04 99fe529
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
