### 2022-10-04 4b5a437
Update builtin_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
