### 2022-10-04 fce56fd
Update custom_macros.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
