+++ doc/gh-pages/notation/custom_macros.md
@@ -7,7 +7,7 @@ nav_order: 10
 # Custom Macros
 
 
-{: .story }   
+{: .story }
 > TODO
 
 
@@ -19,7 +19,7 @@ There are certain system-specific [[Built-In Macros]] for you to take advantage 
 
 Before rolling your dice, you can define some macros
 
-All strings (words) must be capitalized. This is to avoid confusion with the rest of the dice notation 
+All strings (words) must be capitalized. This is to avoid confusion with the rest of the dice notation
 
 `$EVENSIX= {2,4,6,8,10,12}`
 
