### 2022-10-04 d4c70b2
Update dice_counting.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
