+++ doc/gh-pages/notation/dice_counting.md
@@ -30,9 +30,9 @@ nav_order: 8
 In several games, the rolls may not be a matter of *if* your total of rolls is good, but *how many* good rolls you have.
 For this, we count the number of successes, according to a criteria. By default a max-face value is assumed to be a success, but you can change this, much like other condition changes.
 
-> 10d10f>7c - Count how many D10s are larger than 7
+>10d10f>7c - Count how many D10s are larger than 7
 
-> 100dFc - Count how many Fate dice are +
+>100dFc - Count how many Fate dice are +
 
 {: .explanation }
 > Counting dice is a rarer feature of dice rollers and there is at least one instance of using the '#' symbol to denote this operation [[1]](https://www.sophiehoulden.com/dice/documentation/notation.html#count)
@@ -44,7 +44,7 @@ For this, we count the number of successes, according to a criteria. By default 
 
 GNOLL does not support the higher level understanding of the result of dice rolls.
 
-{: .explanation }   
+{: .explanation }
 > Some level of this behaviour can be determined using filters and counts (e.g. count all rolls above 10). However, eeabling this sort of checking within GNOLL would also mean we would have to consider 'degrees of success' and other higher level interpretation of dice rolls, rather than just the rolled result. This would require a large extension of GNOLL's scope.
 >
 > See also the [issue discussion for Degrees of Success](https://github.com/ianfhunter/GNOLL/issues/48)
