### 2022-10-04 c01c387
Update dice_shorthand.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
