+++ doc/gh-pages/notation/dice_shorthand.md
@@ -7,7 +7,7 @@ nav_order: 3
 # Dice Shorthand
 
 
-{: .story }   
+{: .story }
 >**Grindon**: Alright, before I crawl through the chimney, I'm going to look down it to check it's safe
 >
 >**GM**: There are the remains of a fire, but it looks safe to descend
@@ -20,7 +20,7 @@ nav_order: 3
 >
 >**GM**: *Consults Table* "A half-burnt wooden doll that looks eerily familiar. It looks like... you!"
 >
->**Grindon**: That's.. unsettling... I thought my ears were burning earlier.. 
+>**Grindon**: That's.. unsettling... I thought my ears were burning earlier..
 
 ## Numeric Dice
 d% is often used in place of a d100. It is described in some rules as a 'percentile die'
@@ -31,16 +31,16 @@ Instead of a normal N-sided dice, your game may use 'Fate'-style dice. These dic
 Internally these values are treated as -1, 0 and +1.
 
 To use Fate dice, the syntax supports the following:
- - **dF** - a single Fate die
- - **1dF** - a single Fate die
- - **2dF** - multiple Fate die
+- **dF** - a single Fate die
+- **1dF** - a single Fate die
+- **2dF** - multiple Fate die
 
 ## Alternative Fate Dice
 
 Some systems represent Fate dice as U instead of dF. dF is the most popular choice so we have chosen to stick with it.
 Additionally you might find "dF.1", "dF.2" and "dF.3" elsewhere. These expand to:
 
-- dF.3 - [+,+,+,-,-,-] 
+- dF.3 - [+,+,+,-,-,-]
 - dF.2 - [+,+,-,-,0,0] (This is the same as the usual 'dF')
 - dF.1 - [+,0,-,0,0,0]
 
