### 2022-10-04 36ba431
Update drop_keep_notation.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
