### 2022-10-04 a8967b9
Update explosions.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
