### 2022-10-04 20b1295
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
