### 2022-10-04 53b665c
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
