### 2022-10-04 9b57d4f
Update rerolling.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
