### 2022-10-04 13f17c1
Update symbolic_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
