### 2022-10-04 5c0dc6e
Update dice_counting.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
