+++ doc/gh-pages/notation/dice_counting.md
@@ -30,9 +30,11 @@ nav_order: 8
 In several games, the rolls may not be a matter of *if* your total of rolls is good, but *how many* good rolls you have.
 For this, we count the number of successes, according to a criteria. By default a max-face value is assumed to be a success, but you can change this, much like other condition changes.
 
->10d10f>7c - Count how many D10s are larger than 7
+Count how many D10s are larger than 7:
+>10d10f>7c
 
->100dFc - Count how many Fate dice are +
+Count how many Fate dice are '+':
+>100dFc 
 
 {: .explanation }
 > Counting dice is a rarer feature of dice rollers and there is at least one instance of using the '#' symbol to denote this operation [[1]](https://www.sophiehoulden.com/dice/documentation/notation.html#count)
