### 2022-10-04 60c3a24
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
