+++ doc/gh-pages/notation/numeric_dice.md
@@ -39,24 +39,24 @@ What will happen next? [Continue to the next chapter!](Symbolic-Dice)
 
 ## Basic Numeric Dice
 
-The basic form of a dice roll is _x**d**y_ where _x_, _y_ ∈ ℤ<sup>+</sup>.
+The basic form of a dice roll is *x***d***y* where *x*, *y* ∈ ℤ<sup>+</sup>.
 
-In plain english, this means that _x_ and _y_ are positive whole numbers (e.g. 1, 7, 138).
+In plain english, this means that *x* and *y* are positive whole numbers (e.g. 1, 7, 138).
 
-The value _x_ is the number of dice that you wish to roll. You can exclude this, and it'll be assumed you mean a single die.
+The value *x* is the number of dice that you wish to roll. You can exclude this, and it'll be assumed you mean a single die.
 
-The value _y_ tells us how many sides the dice has. A regular six-sided die can be called a "d6". You can't exclude this value, because we won't be able to tell how many sides the dice has!
+The value *y* tells us how many sides the dice has. A regular six-sided die can be called a "d6". You can't exclude this value, because we won't be able to tell how many sides the dice has!
 
 ### Examples
 
 - `1` - A Constant Value. It will just return the value "1"
 - `1d4` - A Single Die Roll with 4 Sides. It will return a value between 1 and 4.
-- `d4` - This is the same as above, as _x_ is assumed to be 1.
+- `d4` - This is the same as above, as **x** is assumed to be 1.
 - `2d4` - Roll 2 Dice with 4 Sides. It will return a value between 2 and 8.
 - `d` - This will return an error because there is no y term.
 
 {: .explanation}
-> The _x_d_y_ notation is one aspect of dice notation that is common for almost every application, so we don't have to make any choices here!
+> The *x***d***y* notation is one aspect of dice notation that is common for almost every application, so we don't have to make any choices here!
 >
 > D without y (as seen in the last example) can default to a set number of sides if a game system only uses one type of dice. However, this default can change between games so as a generic notation we do not assume a value.
 
@@ -85,8 +85,8 @@ Where numbers are all consecutive, one can use a hyphen to indicate a range e.g.
 - d{1,3..6} - produces a value between 1 and 6, but excluding 2
 
 {: .explanation}
-> 
-> With more complicated dice, brings more complicated syntax. However, we want to keep the syntax as straightforward as possible. 
+>
+> With more complicated dice, brings more complicated syntax. However, we want to keep the syntax as straightforward as possible.
 >
 > We note however that there are sometimes easier ways to do these dice if you think about them. e.g. the latter dice could be represented as d6+9
 >
