### 2022-10-04 c7e9c4c
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
