### 2022-10-04 807cf54
Update arithmetic.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
