### 2022-10-04 7a0a191
Update dice_counting.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
