### 2022-10-04 f7216d1
Update numeric_dice.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
