### 2022-10-04 3d3d8f4
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
