+++ doc/gh-pages/etc/supported_rpgs.md
@@ -3,15 +3,17 @@ title: Supported TTRPGs
 published: true
 nav_order: 11
 ---
+
 # {{page.title}}
 
-This is an analysis of all 'notable' TTRPG games on [Wikipedia](https://en.m.wikipedia.org/wiki/List_of_tabletop_role-playing_games) dated 22/Sept/2022. 
+This is an analysis of all 'notable' TTRPG games on [Wikipedia](https://en.m.wikipedia.org/wiki/List_of_tabletop_role-playing_games) dated 22/Sept/2022.
 
 'Support' for a system is exclusive of secondary features that are out of the scope of GNOLL e.g. critical hit notification, degrees of success.
 
 There are a lot of systems here, we cannot guarantee this list is free of error.
 
 ## Status
+
 This list is currently **Under Construction**.
 
 **Letters Done:** `0ABCDQUXYZ`
@@ -22,174 +24,228 @@ This list is currently **Under Construction**.
 
 ## Legend
 
-| Symbol | Meaning |
-| ------ | ------- |
-| ✔️     | Supported |
-| 🔼     | Likely Supported (limited details found on ruleset) |
-| ❓     | Unknown (No details found on ruleset) |
+| Symbol | Meaning                                               |
+| ------ | ----------------------------------------------------- |
+| ✔️     | Supported                                             |
+| 🔼     | Likely Supported (limited details found on ruleset)   |
+| ❓     | Unknown (No details found on ruleset)                 |
 | 🔽     | Unlikely Supported (limited details found on ruleset) |
-| ❌     | Unsupported |
-| 〰️     | Not Applicable |
-|        | Not Evaluted |
+| ❌     | Unsupported                                           |
+| 〰️     | Not Applicable                                        |
+|        | Not Evaluted                                          |
 
 ## List
 
-| TTRPG | Family of Games | Supported | 
-| ----- | ---------- | --------- | 
-| 13th Age [[1]](https://www.13thagesrd.com/combat-rules/) | d20 | ✔️ | 
-| 2300AD |  | 🔼 |
-| 3D&T, 4D&T | d20-like | 🔼 |
-| 7th Sea [[1]](https://www.reddit.com/r/rpg/comments/47j3u8/7th_sea_system/) | Custom | ❓  | 
-| Aberrant | d20/Storyteller | 🔼 |
-| Active Exploits | Diceless | 〰️ | 
-| Advanced Dungeons & Dragons | D&D | ✔️ | 
-| Dungeons & Dragons 3.5 | D&D | ✔️ | 
-| Dungeons & Dragons 4 | D&D | ✔️ | 
-| Dungeons & Dragons 5 | D&D | ✔️ | 
-| Advanced Fighting Fantasy | | ✔️ | 
-| Adventure! (1985) |  | ❓ | 
-| Adventure! (2001) | Modified Storyteller | 🔼 |
-| Adventurer's Guild |  | ❓ | 
-| Adventures in Fantasy [[1]](https://www.scribd.com/document/245240865/1979-Adventures-in-Fantasy-Book-of-Adventure) | |🔼 | 
-| The Adventures of Indiana Jones | | 🔼 |
-| Adventures in Middle Earth | d20 | ✔️ | 
-| After the bomb | Megaversal | ✔️ | 
[…]
