### 2022-10-04 334277c
generalize makefile
