+++ Makefile
@@ -1,5 +1,26 @@
+CODEDIRS=./src/grammar ./src/grammar/rolls
+INCDIRS=./src/grammar
 
-.PHONY: clean yacc lex compile all test test_no_pip
+CC=cc
+OPT=-O3
+
+# add flags and the include paths
+CFLAGS=$(foreach D,$(INCDIRS),-I$(D)) $(OPT)
+
+# add flags to build for shared library and add include paths
+SHAREDCFLAGS=-fPIC -c $(foreach D,$(INCDIRS),-I$(D))
+
+# generate list of c files and remove y.tab.c from src/grammar directory
+CFILES=$(foreach D,$(CODEDIRS),$(wildcard $(D)/*.c)) build/lex.yy.c build/y.tab.c
+CFILES:=$(filter-out ./src/grammar/y.tab.c, $(CFILES))
+
+# list out object files from the above c files. Replace their path with build/
+OBJECTS=$(addprefix build/,$(notdir $(patsubst %.c,%.o,$(CFILES))))
+
+
+all: clean yacc lex compile shared
+	echo "== Build Complete =="
+
 yacc:
 	mkdir -p build
 	if [ -z $(DEBUG) ]; then \
@@ -15,34 +36,37 @@ lex:
 	lex src/grammar/dice.lex
 	mv lex.yy.c build/lex.yy.c
 
-compile:
 # Executable
-	cc -O3 build/y.tab.c \
-		src/grammar/rolls/sided_dice.c \
-		src/grammar/rolls/condition_checking.c \
-		src/grammar/vector_functions.c \
-		src/grammar/dice_logic.c \
-		src/grammar/macro_logic.c \
-		build/lex.yy.c \
-		-Isrc/grammar/
+compile:
+	$(CC) $(CFLAGS) $(CFILES)
 
 # Shared Lib
-	cc -fPIC -c build/y.tab.c -o build/tab.o -Isrc/grammar/
-	cc -fPIC -c src/grammar/vector_functions.c -o build/vec.o -Isrc/grammar/
-	cc -fPIC -c src/grammar/dice_logic.c -o build/die.o -Isrc/grammar/
-	cc -fPIC -c src/grammar/macro_logic.c -o build/macro.o -Isrc/grammar/
-	cc -fPIC -c src/grammar/rolls/sided_dice.c -o build/rso.o -Isrc/grammar/
-	cc -fPIC -c src/grammar/rolls/condition_checking.c -o build/cc.o -Isrc/grammar/
-	cc -fPIC -c build/lex.yy.c -o build/lex.o  -Isrc/grammar/
-	cc -shared -o build/dice.so build/die.o build/macro.o build/tab.o build/cc.o build/lex.o build/vec.o build/rso.o
+shared: $(OBJECTS)
+	$(CC) -shared -o build/$@ $^
 
 # Linux
 	mv ./a.out build/dice | true
 # Windows
 	mv ./a.exe build/dice | true
 
-all: clean yacc lex compile
-	echo "== Build Complete =="
+# hardcode for y.tab.o
+build/y.tab.o: 
+	$(CC) $(SHAREDCFLAGS) -c build/y.tab.c -o $@
+
+build/lex.yy.o:
+	$(CC) $(SHAREDCFLAGS) -c build/lex.yy.c -o $@
+
+# for /grammar/rolls hardcode
+build/condition_checking.o:
+	$(CC) $(SHAREDCFLAGS) -c src/grammar/rolls/condition_checking.c -o $@
+
+# for /grammar/rolls hardcode
+build/sided_dice.o:
[…]
