### 2022-10-04 726848f
fix: add new target in go/target.mk and fix .so name
