+++ .github/workflows/label.yml
@@ -6,14 +6,13 @@
 # https://github.com/actions/labeler
 
 name: "Automation: PR Labels"
-on: [pull_request]
 
+on:
+  - pull_request_target
 jobs:
   label:
-
     runs-on: ubuntu-latest
     permissions:
-      contents: read
       pull-requests: write
 
     steps:
