+++ Makefile
@@ -55,6 +55,7 @@ include src/js/target.mk
 include src/go/target.mk
 include src/perl/target.mk
 include src/swig/target.mk
+include src/haskell/target.mk
 
 clean: perl_clean python_clean
 	rm -rf build
