### 2022-10-05 4b8b297
add functions to syntax (placeholder)
