+++ src/grammar/vector_functions.h
@@ -11,6 +11,7 @@ void pop(int * arr, int len, int value, int * new_arr);
 
 int min(int * arr, int len);
 int max(int * arr, int len);
+void abs_vec(vec * x);
 
 void concat_symbols(char ** arr1, int len1,
                     char ** arr2, int len2,
