+++ doc/gh-pages/etc/supported_rpgs.md
@@ -173,15 +173,15 @@ This list is currently **Under Construction**.
 | DC adventures                                                                                                               | d20                     | ✔️                                                          |
 | DC Heroes                                                                                                                   | MEGS                    | ✔️                                                          |
 | DC Universe                                                                                                                 | d6                      | ✔️                                                          |
-| Dead inside                                                                                                                 | PDQ                     | ❓                                                          |
+| Dead inside                                                                                                                 | PDQ                     | 🔼                                                          |
 | De Profundis                                                                                                                |                         | 〰️                                                          |
-| Dead Inside                                                                                                                 |                         |                                                             |
-| Dead Reign                                                                                                                  |                         |                                                             |
-| Deadlands: Hell on Earth                                                                                                    |                         |                                                             |
-| Deadlands: Lost Colony                                                                                                      |                         |                                                             |
+| Dead Inside                                                                                                                 | PDQ                     | 🔼                                                          |
+| Dead Reign                                                                                                                  |                         | ❓                                                          |
+| Deadlands: Hell on Earth                                                                                                    | d20/Savage Worlds       | 🔼                                                          |
+| Deadlands: Lost Colony                                                                                                      |                         | ❓                                                          |
 | Deadlands: The Weird West                                                                                                   | GURPS                   | ✔️                                                          |
 | Deathwatch                                                                                                                  |                         |                                                             |
-| Deliria                                                                                                                     |                         |                                                             |
+| Deliria                                                                                                                     | Compact                 | ❓                                                          |
 | Delta Force: America Strikes Back!                                                                                          |                         |                                                             |
 | Degenesis: Rebirth                                                                                                          | Katharsys               |                                                             |
 | Demon City Shinjuku Role-Playing Game                                                                                       |                         |                                                             |
