### 2022-10-06 9b441b6
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
