+++ doc/gh-pages/etc/supported_rpgs.md
@@ -180,17 +180,17 @@ This list is currently **Under Construction**.
 | Deadlands: Hell on Earth                                                                                                    | d20/Savage Worlds       | 🔼                                                          |
 | Deadlands: Lost Colony                                                                                                      |                         | ❓                                                          |
 | Deadlands: The Weird West                                                                                                   | GURPS                   | ✔️                                                          |
-| Deathwatch                                                                                                                  |                         |                                                             |
+| Deathwatch                                                                                                                  |                         | ✔️                                                          |
 | Deliria                                                                                                                     | Compact                 | ❓                                                          |
-| Delta Force: America Strikes Back!                                                                                          |                         |                                                             |
-| Degenesis: Rebirth                                                                                                          | Katharsys               |                                                             |
-| Demon City Shinjuku Role-Playing Game                                                                                       |                         |                                                             |
-| Demon Hunters Role Playing Game                                                                                             |                         |                                                             |
-| Demon: The Fallen                                                                                                           | World of Darkness       |                                                             |
-| Demon's Lair                                                                                                                |                         |                                                             |
-| Deryni                                                                                                                      |                         |                                                             |
-| Diana: Warrior Princess                                                                                                     |                         |                                                             |
-| Diaspora VSCA                                                                                                               | Fate                    |                                                             |
+| Delta Force: America Strikes Back!                                                                                          |                         | ❓                                                           |
+| Degenesis: Rebirth                                                                                                          | Katharsys               | ✔️                                                         |
+| Demon City Shinjuku Role-Playing Game                                                                                       | BESM                    | ✔️                                                            |
+| Demon Hunters Role Playing Game                                                                                             | Storyteller             | ✔️                                                          |
+| Demon: The Fallen                                                                                                           | World of Darkness       | ✔️                                                          |
+| Demon's Lair                                                                                                                |                         | ✔️                                                          |
+| Deryni                                                                                                                      | Fudge                   | ✔️                                                          |
+| Diana: Warrior Princess                                                                                                     | Fate                    | ✔️                                                          |
+| Diaspora VSCA                                                                                                               | Fate                    | ✔️                                                         |
 | Dinky Dungeons                                                                                                              |                         |                                                             |
 | Dinosaur Planet: Broncosaurus Rex                                                                                           |                         |                                                             |
 | Diomin                                                                                                                      |                         |                                                             |
@@ -205,7 +205,7 @@ This list is currently **Under Construction**.
 | Donjon Clinton                                                                                                              | Donjon                  |                                                             |
 | Don't Look Back                                                                                                             |                         |                                                             |
 | Don't Rest Your Head                                                                                                        |                         |                                                             |
-| Drací doupe                                                                                                                 | D&D-esqe                |                                                             |
+| Drací doupe                                                                                                                 | D&D-esqe                | ✔️                                                          |
 | Dragon Age                                                                                                                  |                         |                                                             |
 | Dragon Ball Z                                                                                                               |                         |                                                             |
 | Dragonlance: Fifth Age                                                                                                      |                         |                                                             |
@@ -231,8 +231,8 @@ This list is currently **Under Construction**.
 | Dungeon World                                                                                                               | PbtA                    | ✔️                                                          |
 | Durance                                                                                                                     |                         |                                                             |
 | Dust Devils                                                                                                                 |                         |                                                             |
-| Dying Earth                                                                                                                 |                         |                                                             |
-| Dzikie Pola (Wild Fields)                                                                                                   |                         |                                                             |
+| Dying Earth                                                                                                                 |                         | ✔️                                                            |
+| Dzikie Pola (Wild Fields)                                                                                                   |                         | ✔️                                                          |
[…]
