### 2022-10-06 746fb2d
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
