+++ .github/workflows/js-test.yml
@@ -21,6 +21,7 @@ jobs:
       run: python3 -m pip install -r reqs/requirements.txt
     - name: javascript prerequisites
       run: |
+        sudo apt install clang-14 lld-14 llvm-14 nodejs libjs-d3 python3-numpy
         sudo apt install nodejs llvm -y
         wget -c http://archive.ubuntu.com/ubuntu/pool/universe/e/emscripten/emscripten_3.1.6~dfsg-5_all.deb
         sudo apt install ./emscripten_3.1.6~dfsg-5_all.deb
