### 2022-10-06 97fe9c5
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
