### 2022-10-06 db4d44a
Update js-test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
