+++ SECURITY.md
@@ -1,0 +1,12 @@
+# Security Policy
+
+## Supported Versions
+
+At this time, only the latest version of GNOLL is supported for security updates.
+
+## Reporting a Vulnerability
+
+To report a vulnerability, either log an issue in our Github issue tracker.
+Only if it is high priority should authors/developers be mailed directly.
+
+We will work on this issue depending on developer availability and severity. Please appreciate that all contributors to GNOLL are unpaid volunteers.
