### 2022-10-06 501828a
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
