### 2022-10-06 9d478b5
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
