+++ doc/gh-pages/etc/supported_rpgs.md
@@ -136,7 +136,7 @@ This list is currently **Under Construction**.
 | Corporation                                                                                                                 | Brutal                  | ✔️                                                          |
 | CORPS [[1]](https://www.drivethrurpg.com/m/product/1098)                                                                    |                         | ✔️                                                          |
 | Corum                                                                                                                       | Elric!                  | 🔼                                                          |
-| Creeks and Crawdads                                                                                                         | ✔️                      |                                                             |
+| Creeks and Crawdads                                                                                                         |                      |    ✔️                                                          |
 | Crime Fighter [[1]](https://dmmagazine.blogspot.com/2022/02/analysis-crime-fighter-role-playing.html?m=1)                   |                         | 🔼                                                          |
 | Crimefighters                                                                                                               |                         | ❓                                                          |
 | Crystalicum                                                                                                                 |                         | ❓                                                          |
@@ -198,24 +198,24 @@ This list is currently **Under Construction**.
 | The Doctor Who Role Playing Game                                                                                            | Vortex                  |  ✔️                                                        |
 | Doctor Who - Adventures In Time And Space                                                                                   |                         | 🔼                                                         |
 | Doctor Who - Time Lord                                                                                                      |                         | 🔼                                                         |
-| Dogs in the Vineyard                                                                                                        |                         | 🔼 *                                                           |
-| Domination                                                                                                                  |                         | ❓                                                            |
-| Dominion Rules                                                                                                              |                         | ✔️                                                           |
-| The Dominion Tank Police                                                                                                    | Tri-Stat                | ✔️                                                          |
-| Donjon                                                                                                                      | Donjon                  | 🔼                                                          |
-| Don't Look Back                                                                                                             |                         |                                                             |
-| Don't Rest Your Head                                                                                                        |                         |                                                             |
-| Drací doupe                                                                                                                 | D&D-esqe                | ✔️                                                          |
-| Dragon Age                                                                                                                  |                         |                                                             |
-| Dragon Ball Z                                                                                                               |                         |                                                             |
-| Dragonlance: Fifth Age                                                                                                      |                         |                                                             |
-| Dragon Lords of Melniboné                                                                                                   |                         |                                                             |
-| DragonMech                                                                                                                  |                         |                                                             |
-| DragonQuest                                                                                                                 |                         |                                                             |
-| DragonRaid                                                                                                                  |                         |                                                             |
-| Dragonroar                                                                                                                  |                         |                                                             |
-| Dragonstar                                                                                                                  |                         |                                                             |
-| Dragon Storm                                                                                                                |                         |                                                             |
+| Dogs in the Vineyard                                                                                                        |                         | 🔼 *                                                       |
+| Domination                                                                                                                  |                         | ❓                                                         |
+| Dominion Rules                                                                                                              |                         | ✔️                                                         |
+| The Dominion Tank Police                                                                                                    | Tri-Stat                | ✔️                                                         |
+| Donjon                                                                                                                      | Donjon                  | 🔼                                                         |
+| Don't Look Back                                                                                                             |                         | ❓                                                         |
+| Don't Rest Your Head                                                                                                        |                         | ❌                                                         |
+| Drací doupe                                                                                                                 | D&D-esqe                | ✔️                                                         |
+| Dragon Age                                                                                                                  |                         | 🔼                                                         |
+| Dragon Ball Z                                                                                                               |                         | ❓                                                            |
+| Dragonlance: Fifth Age                                                                                                      | SAGA                    | 〰️                                                             |
+| Dragon Lords of Melniboné                                                                                                   | d20                     | 🔼                                                            |
+| DragonMech                                                                                                                  | d20                     | 🔼                                                          |
+| DragonQuest                                                                                                                 |                         |  ✔️                                                            |
[…]
