### 2022-10-08 4523e12
Correct display of the story.
