### 2022-10-08 2536b1d
Write the story for builtin macros.
