+++ doc/gh-pages/notation/builtin_macros.md
@@ -8,7 +8,23 @@ nav_order: 9
 # Built-In Macros
 
 {: .story }
-> TODO
+>**GM**: You fail to secure yourself on the wall and slip towards the spikes. While falling, you hit your head against the wall. 
+>
+>**Grindon**: Gosh, that hurts.
+>
+>**GM**: You miraculously hit the ground without dying. But your blow to the head makes you delirious.
+>
+>**Grindon**: Oh look, fairies wearing armors painted with Zodiac signs!
+>
+>**GM**: Yes... (sighs) As they fligh around your head, you decide to eat one of them.
+>
+>**Grindon**: Okay that's #PLANETS.
+>
+>*GNOLL*: [Mercury through Pluto]
+>
+>**Grindon The Brave**: Well... not my favorite taste.
+>
+>**GM**: Entirely understandable. But that is not a 1, so no reroll. But don't worry, something better is coming.
 
 ## Supported Macros
 Here are a sample of the dice that can be used via Macro Accessors
