### 2022-10-08 d514858
Reword one sentence.
