### 2022-10-08 5f6e225
Write the story for custom macros.
