+++ doc/gh-pages/notation/custom_macros.md
@@ -6,10 +6,22 @@ nav_order: 10
 
 # Custom Macros
 
-
 {: .story }
-> TODO
-
+>**GM**: You wake up from your dream, finding an old creature resembling to a witch. Seeing your bleeding head, it offers you to choose a potion.
+>
+>**Grindon**: Can I pick the healing potion?
+>
+>**GM**: Almost. There is one bright-green healing potion. There is also one deep-red deadly potion, that would also help with your bleeding head, but not with your quest. And finally an intriguing blue potion, full of sprinkles swirling inside the bottle.
+>
+>**Grindon**: Ok. How do I pick one?
+>
+>**GM**: Extreme solutions like this one require special die. This one is a Please-Save-Me die, with 3 faces named "life", "death" and "something else".
+>
+>**Grindon**: Okay! Com'on die.
+>
+>*GNOLL*: [something else]
+>
+>**Grindon The Brave**: I am starting to fell weird...
 
 ## Usage
 
