### 2022-10-09 ianfhunter/GNOLL#249 @Smattr
validate_roll does not check number of dice
No problem, just something I noticed while compiling the code.

**Describe the solution you'd like**

It seems like this function should check the number of dice is non-negative.

None.

**Additional context**

None.
