### 2022-10-09 ianfhunter/GNOLL#251 @Smattr
missing error handling
**Describe the bug**

There are a number of places in the code where a C API that can fail is called and failure is not checked for. The outcome of such failures will likely be a crash, something that is undesirable when GNOLL is called as a library by a containing application that may want to be robust against such things. In the context of a cross-language binding, a failure like this when reached e.g. from Python, will crash the containing Python interpreter.

**To Reproduce**

The most straightforward instance is memory allocations, e.g. src/grammar/macro_logic.c:15:16:

```c
s = (struct macro_struct*)malloc(sizeof *s);
s->id = key;
```

On failure, `malloc` returns `NULL`, which will then result in the second line above dereferencing a null pointer.

A different example is `fopen`, src/grammar/dice.yacc:133

```c
fp = fopen(output_file, "a+");
…
fprintf(fp, …);
```

This will again lead to a null pointer dereference if the file opening fails.

**Expected behavior**

In these cases, tests that the return value is non-null. If it's null then some way of indicating failure to the caller will be needed. For `fopen`, this may require examining `errno` to figure out exactly what went wrong.
