### 2022-10-09 ianfhunter/GNOLL#252 @Smattr
set a C standard in the Makefile
The Makefile does not seem to set any ISO C standard with which to compile to code (`/std` to MSVC, `-std=…` to other compilers). Without this, the standard used during compilation will be the compiler's default. This varies across e.g. GCC releases.

While it's unusual, versions of the C standard can and do make breaking changes. So leaving the C standard unspecified is likely to lead to “works on my machine” kind of problems, where users get code with differing semantics depending on the compiler in use at build time.

**Describe the solution you'd like**

I would suggest picking the latest practically available standard. This generally means C11. However, it you support users compiling with MSVC, you'll probably need to pick C99 as (AFAIK) it is the latest one MSVC supports.

Something like:

```diff
diff --git Makefile Makefile
index 11e7f56..d2b78cc 100644
--- Makefile
+++ Makefile
@@ -2,7 +2,7 @@ CODEDIRS=./src/grammar ./src/grammar/rolls
 INCDIRS=./src/grammar

 CC=cc
-OPT=-O3
+OPT=-O3 -std=c11

 # add flags and the include paths
 CFLAGS=$(foreach D,$(INCDIRS),-I$(D)) $(OPT)
```

None.

**Additional context**

Please let me know if you have questions or if what I wrote was unclear.
