### 2022-10-09 f76dd1c
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
