+++ src/grammar/dice_logic.c
@@ -144,6 +144,9 @@ int validate_roll(int number_of_dice, int die_side)
     if (die_side < 0){
         printf("Cannot roll a dice with a negative amount of sides\n");
         return 1;
+    }if (number_of_dice < 0){
+        printf("Cannot roll a negative number of dice\n");
+        return 1;
     }
     return 0;
 }
