### 2022-10-09 ianfhunter/GNOLL#252 @ianfhunter
Totally agree, thank you for flagging.

In #253 I added c99 as the standard. 👍
