### 2022-10-09 1bfb806
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
