### 2022-10-09 470e9a4
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
