### 2022-10-09 4f72b69
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
