### 2022-10-09 96a96d5
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
