### 2022-10-09 9b54516
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
