+++ src/grammar/macro_logic.c
@@ -2,6 +2,7 @@
 #include "uthash.h"
 #include "vector_functions.h"
 #include "macro_logic.h"
+#include "shared_header.h"
 
 struct macro_struct *macros = NULL; //Initialized to NULL (Importnat)
 
@@ -13,6 +14,9 @@ void register_macro(char * skey, vec *to_store) {
     HASH_FIND_INT(macros, &key, s);  /* id already in the hash? */
     if (s == NULL){
         s = (struct macro_struct*)malloc(sizeof *s);
+        if(! vector->content){
+            exit(BAD_ALLOC);
+        }
         s->id = key;
         HASH_ADD_INT(macros, id, s);  /* id: name of key field */
     }
