### 2022-10-09 901a9a8
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
