### 2022-10-09 8c6580b
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
