### 2022-10-09 65ed18b
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
