### 2022-10-09 00879b3
Create safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
