### 2022-10-09 d239a9a
Create safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
