### 2022-10-09 310306c
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
