### 2022-10-09 2b97f44
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
