### 2022-10-09 a181822
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
