+++ src/grammar/shared_header.h
@@ -17,7 +17,8 @@ typedef enum {
     BAD_ALLOC=1,
     BAD_FILE=2,
     NOT_IMPLEMENTED=3,
-    INTERNAL_ASSERT=4
+    INTERNAL_ASSERT=4,
+    UNDEFINED_BEHAVIOUR=5
 } ERROR_CODES;
 
 int roll(char * s);
