### 2022-10-09 2d0c086
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
