### 2022-10-09 9673d6c
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
