### 2022-10-09 af79b89
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
