### 2022-10-09 6ea7be3
Update safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
