### 2022-10-09 27f2406
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
