+++ src/grammar/safe_functions.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "shared_header.h"
+#include <stdarg.h>
 
 void * safe_malloc(size_t size){
     void * malloc_result;
@@ -20,6 +21,7 @@ void * safe_calloc(size_t nitems, size_t size){
     }
     return calloc_result;
 }
+
 FILE * safe_fopen(const char *filename, const char *mode){
     FILE * fopen_result;
     fopen_result = fopen(filename, mode);
@@ -38,7 +40,6 @@ char * safe_strdup( const char *str1 ){
     return result;
 }
 
-
 long int safe_strtol (const char* str, char** endptr, int base){
     long int result;
     result = strtol(str,endptr,base);
@@ -46,4 +47,22 @@ long int safe_strtol (const char* str, char** endptr, int base){
         exit(OUT_OF_RANGE);
     }
     return result;
+}
+
+void safe_printf(const char *fmt, ...) {
+    va_list args;
+    va_start(args, fmt);
+    int count = vprintf(fmt, args);
+    va_end(args);
+    if(count < 0) exit(IO_ERROR);
+    return count
+}
+
+void safe_fprintf(FILE *stream, const char *format, ...) {
+    va_list args;
+    va_start(args, fmt);
+    int count = vfprintf(stream, format, args);
+    va_end(args);
+    if(count < 0) exit(IO_ERROR);
+    return count
 }
