### 2022-10-09 75cb005
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
