### 2022-10-09 1a29e47
Update safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
