### 2022-10-09 9dbfd2f
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
