### 2022-10-09 f6c7500
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
