### 2022-10-09 c912fa0
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
