+++ src/grammar/safe_functions.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdarg.h>
+#include <errno.h>
 #include "shared_header.h"
 #include "safe_functions.h"
-#include <stdarg.h>
 
 void * safe_malloc(size_t size){
     void * malloc_result;
