### 2022-10-09 5c96cc7
Update safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
