### 2022-10-09 50ada8b
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
