### 2022-10-09 1bbbecb
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
