### 2022-10-09 8db1631
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
