### 2022-10-09 a5175f7
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
