### 2022-10-09 a97fee8
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
