### 2022-10-09 5789026
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
