### 2022-10-09 226c4f1
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
