### 2022-10-09 e17b405
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
