### 2022-10-09 88b0f92
Update util.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
