### 2022-10-09 1fe54e1
Update util.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
