### 2022-10-09 a75d40f
Update util.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
