+++ tests/python/util.py
@@ -49,7 +49,7 @@ def roll(s, mock_mode=Mock.NO_MOCK, mock_const=3):
         mock_mode = Mock.NO_MOCK
     first_run = False
 
-    print(f"Test Roll: '{s}')
+    print(f"Test Roll: '{s}'")
 
     # Get module now - post make
     dice_tower_roll = get_roll()
