### 2022-10-09 4319832
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
