### 2022-10-09 98836cf
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
