+++ src/python/code/gnoll/parser.py
@@ -3,6 +3,7 @@ import sys
 import tempfile
 
 import cppyy
+from enum import Enum
 
 BUILD_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "c_build"))
 C_HEADER = os.path.join(os.path.dirname(__file__), "c_includes")
@@ -12,6 +13,24 @@ cppyy.c_include(os.path.join(C_HEADER, "shared_header.h"))
 cppyy.c_include(os.path.join(C_HEADER, "dice_logic.h"))
 cppyy.load_library(C_SHARED_LIB)
 
+
+class GNOLLException(Exception):
+    def __init__(self, v):
+        Exception.__init__(self, 'Tax ID expired')
+
+def RaiseGNOLLError(v):
+    d = [
+        None,
+        GNOLLException("BAD_ALLOC"),
+        GNOLLException("BAD_FILE"),
+        GNOLLException("NOT_IMPLEMENTED"),
+        GNOLLException("INTERNAL_ASSERT"),
+        GNOLLException("UNDEFINED_BEHAVIOUR"),
+        GNOLLException("BAD_STRING"),
+        GNOLLException("OUT_OF_RANGE"),
+        GNOLLException("IO_ERROR")
+    ]
+    raise d[v]
 
 def roll(s, verbose=False, mock=None, quiet=True, mock_const=3):     
     if verbose:
@@ -29,14 +48,13 @@ def roll(s, verbose=False, mock=None, quiet=True, mock_const=3):
 
     cppyy.gbl.reset_mocking()
     if mock is None:
-        print("A")
         return_code = cppyy.gbl.roll_and_write(s, f)
     else:
         # Testing Only
-        print("B")
         return_code = cppyy.gbl.mock_roll(s, f, mock, quiet, mock_const)
-
-    print("Return Code:", return_code)
+    
+    if(return_code != 0):
+        RaiseGNOLLError(return_code)
 
     if verbose:
         print("Temp File:", temp.name)
