### 2022-10-09 a63e0bf
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
