### 2022-10-09 821cb3d
Update parser.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
