### 2022-10-09 dda9368
Update util.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
