### 2022-10-09 0ac439c
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
