### 2022-10-09 13c5140
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
