### 2022-10-09 cdb1105
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
