+++ src/grammar/safe_functions.c
@@ -36,7 +36,7 @@ FILE * safe_fopen(const char *filename, const char *mode){
 
 int safe_fclose(FILE *stream){
     if(fclose(stream) != 0){
-        printed("err closing\n");
+        printf("err closing\n");
         exit(BAD_FILE);
     }
     return 0;
