### 2022-10-09 2c4a4cf
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
