### 2022-10-09 6d737c1
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
