+++ src/grammar/safe_functions.c
@@ -6,11 +6,13 @@
 #include "shared_header.h"
 #include "safe_functions.h"
 
+unsigned int gnoll_errno = 0;
+
 void * safe_malloc(size_t size){
     void * malloc_result;
     malloc_result = malloc(size);
     if(!malloc_result){
-        exit(BAD_ALLOC);
+        gnoll_errno = BAD_ALLOC;
     }
     return malloc_result;
 }
@@ -19,7 +21,7 @@ void * safe_calloc(size_t nitems, size_t size){
     void * calloc_result;
     calloc_result = calloc(nitems, size);
     if(!calloc_result){
-        exit(BAD_ALLOC);
+        gnoll_errno = BAD_ALLOC;
     }
     return calloc_result;
 }
@@ -30,7 +32,7 @@ FILE * safe_fopen(const char *filename, const char *mode){
     if(fopen_result == NULL){
         printf("err opening\n");
         perror(filename);
-        exit(BAD_FILE);
+        gnoll_errno = BAD_FILE;
     }
     return fopen_result;
 }
@@ -38,7 +40,7 @@ FILE * safe_fopen(const char *filename, const char *mode){
 int safe_fclose(FILE *stream){
     if(fclose(stream) != 0){
         printf("err closing\n");
-        exit(BAD_FILE);
+        gnoll_errno = BAD_FILE;
     }
     return 0;
 }
@@ -49,7 +51,7 @@ char * safe_strdup( const char *str1 ){
     result = safe_calloc(sizeof(char), l);
     result = strcpy(result, str1);
     if(result != 0){
-        exit(BAD_STRING);
+        gnoll_errno = BAD_STRING;
     }
     return result;
 }
@@ -58,7 +60,7 @@ long int safe_strtol (const char* str, char** endptr, int base){
     long int result;
     result = strtol(str,endptr,base);
     if(errno == ERANGE){
-        exit(OUT_OF_RANGE);
+        gnoll_errno = OUT_OF_RANGE;
     }
     return result;
 }
@@ -68,7 +70,7 @@ void safe_printf(const char *fmt, ...) {
     va_start(args, fmt);
     int count = vprintf(fmt, args);
     va_end(args);
-    if(count < 0) exit(IO_ERROR);
+    if(count < 0) gnoll_errno = IO_ERROR;
 }
 
 void safe_fprintf(FILE *stream, const char *fmt, ...) {
@@ -76,5 +78,5 @@ void safe_fprintf(FILE *stream, const char *fmt, ...) {
     va_start(args, fmt);
     int count = vfprintf(stream, fmt, args);
     va_end(args);
-    if(count < 0) exit(IO_ERROR);
+    if(count < 0) gnoll_errno = IO_ERROR;
 }
