### 2022-10-10 077029d
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
