### 2022-10-10 9030c08
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
