### 2022-10-10 41fe4e5
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
