### 2022-10-10 49f7f48
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
