### 2022-10-10 731a995
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
