### 2022-10-10 8e47a05
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
