### 2022-10-10 e649d38
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
