### 2022-10-10 2850632
Update test_filtering.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
