### 2022-10-10 a07840f
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
