+++ src/grammar/safe_functions.c
@@ -50,7 +50,7 @@ char * safe_strdup( const char *str1 ){
     unsigned int l = strlen(str1);
     result = safe_calloc(sizeof(char), l);
     result = strcpy(result, str1);
-    if(result != 0){
+    if(result == 0){
         gnoll_errno = BAD_STRING;
     }
     return result;
