### 2022-10-10 bf83b56
Update condition_checking.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
