### 2022-10-10 93243dc
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
