### 2022-10-10 25fc1ba
Update test_filtering.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
