### 2022-10-10 e875c8e
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
