### 2022-10-10 fd97816
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
