### 2022-10-10 b9b7de6
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
