+++ src/grammar/macro_logic.c
@@ -8,7 +8,11 @@
 struct macro_struct *macros = NULL; //Initialized to NULL (Importnat)
 extern unsigned int gnoll_errno;
 
-void register_macro(char * skey, vec *to_store) {
+
+void register_macro(vec * macro_name, vec *to_store) {
+    if(gnoll_errno) return;
+
+    char * skey = macro_name->symbols[0];
     int key = safe_strtol(skey, NULL, 10);
     if(gnoll_errno) return;
 
@@ -21,10 +25,12 @@ void register_macro(char * skey, vec *to_store) {
         s->id = key;
         HASH_ADD_INT(macros, id, s);  /* id: name of key field */
     }
-    memcpy(&s->stored_dice_roll, &to_store, sizeof(to_store));
+    memcpy(&s->stored_dice_roll, &to_store, sizeof(*to_store));
 }
 
 struct macro_struct *search_macros(char * skey, vec *to_store) {
+    if(gnoll_errno) return NULL;
+
     int key = safe_strtol(skey, NULL, 10);
     if(gnoll_errno) return NULL;
 
