+++ src/grammar/safe_functions.c
@@ -9,6 +9,11 @@
 unsigned int gnoll_errno = 0;
 
 void * safe_malloc(size_t size){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return NULL;
+    }
     void * malloc_result;
     malloc_result = malloc(size);
     if(!malloc_result){
@@ -18,6 +23,11 @@ void * safe_malloc(size_t size){
 }
 
 void * safe_calloc(size_t nitems, size_t size){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return NULL;
+    }
     void * calloc_result;
     calloc_result = calloc(nitems, size);
     if(!calloc_result){
@@ -27,6 +37,11 @@ void * safe_calloc(size_t nitems, size_t size){
 }
 
 FILE * safe_fopen(const char *filename, const char *mode){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return NULL;
+    }
     FILE * fopen_result;
     fopen_result = fopen(filename, mode);
     if(fopen_result == NULL){
@@ -38,6 +53,11 @@ FILE * safe_fopen(const char *filename, const char *mode){
 }
 
 int safe_fclose(FILE *stream){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return 0;
+    }
     if(fclose(stream) != 0){
         printf("err closing\n");
         gnoll_errno = BAD_FILE;
@@ -46,6 +66,11 @@ int safe_fclose(FILE *stream){
 }
 
 char * safe_strdup( const char *str1 ){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return NULL;
+    }
     char * result;
     unsigned int l = strlen(str1);
     result = safe_calloc(sizeof(char), l);
@@ -57,6 +82,11 @@ char * safe_strdup( const char *str1 ){
 }
 
 long int safe_strtol (const char* str, char** endptr, int base){
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return 0;
+    }
     long int result;
     result = strtol(str,endptr,base);
     if(errno == ERANGE){
@@ -66,6 +96,11 @@ long int safe_strtol (const char* str, char** endptr, int base){
 }
 
 void safe_printf(const char *fmt, ...) {
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return;
+    }
     va_list args;
     va_start(args, fmt);
     int count = vprintf(fmt, args);
@@ -74,6 +109,11 @@ void safe_printf(const char *fmt, ...) {
 }
 
 void safe_fprintf(FILE *stream, const char *fmt, ...) {
+    if (gnoll_errno){
+        // If there was already an error,
+        // Don't even try to execute.
+        return;
+    }
     va_list args;
     va_start(args, fmt);
     int count = vfprintf(stream, fmt, args);
