### 2022-10-10 8869b0c
move js artifacts to build
