+++ tests/python/test_macros.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import pytest
-from util import roll, Mock
+from util import Mock, roll
 
 @pytest.mark.parametrize("r,out,mock",[
     ("#MY_DIE=d{A};d4", 3, Mock.RETURN_CONSTANT),
