+++ tests/python/util.py
@@ -1,7 +1,7 @@
+from enum import Enum
 import importlib.util as iu
 import os
 import subprocess
-from enum import Enum
 
 GRAMMAR_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../src/grammar"))
 SRC_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../src/python/code/gnoll/"))
