### 2022-10-10 308d3b2
Update setup.cfg
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
