### 2022-10-10 872fb37
Update supported_rpgs.md
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
