+++ doc/gh-pages/etc/supported_rpgs.md
@@ -212,15 +212,15 @@ This list is currently **Under Construction**.
 | Dragon Lords of Melniboné                                                                                                   | d20                     | 🔼                                                            |
 | DragonMech                                                                                                                  | d20                     | 🔼                                                          |
 | DragonQuest                                                                                                                 |                         |  ✔️                                                            |
-| DragonRaid                                                                                                                  | ALS                   |  ✔️                                                           |
+| DragonRaid                                                                                                                  | ALS                    |  ✔️                                                           |
 | Dragonroar                                                                                                                  |                         |   ❓                                                          |
 | Dragonstar                                                                                                                  |                         |   ❓                                                          |
 | Dragon Storm                                                                                                                |                         |  ❓                                                           |
-| Dragon Warriors                                                                                                             |                         |                                                             |
-| Drakar och Demoner (Dragons and Demons)                                                                                     |                         |                                                             |
-| Dread RPG                                                                                                                   |                         |                                                             |
-| Dread                                                                                                                       |                         |                                                             |
-| Dream Park                                                                                                                  |                         |                                                             |
+| Dragon Warriors                                                                                                             |                         |   ✔️                                                          |
+| Drakar och Demoner (Dragons and Demons)                                                                                     |                         |   🔼                                                          |
+| Dread RPG                                                                                                                   |                         |    〰️                                                         |
+| Dread                                                                                                                       |                         |     🔼                                                          |
+| Dream Park                                                                                                                  |                         |     🔼                                                          |
 | The Dresden Files                                                                                                           | FATE                    | ✔️                                                          |
 | Droids                                                                                                                      |                         |  ❓                                                           |
 | Dune: Chronicles of the Imperium                                                                                            |                         |  🔼                                                        |
