### 2022-10-10 #4816 @galessiorob
👋 @ziotom78, please create your checklist by typing: `@editorialbot generate my checklist`

This will give you an itemized list for you to follow. Please let me know if you have any questions. Thanks!
