### 2022-10-10 1c4cee9
add working roll_and_write func
