### 2022-10-11 ianfhunter/GNOLL#251 @Smattr
Wow, speedy! I didn't expect you to address this overnight! You must type much faster than I do.

A few thoughts:
* You might want to check the size during `safe_malloc`/`safe_calloc` before considering a `NULL` return to be an error. It's technically permitted for these functions to succeed and return `NULL` if any of their inputs are `0`. So `NULL` only indicates failure when all inputs are `>0`. As some trivia, I think Glibc's allocator returns non-null for 0-sized allocations due to some legacy assumptions in Emacs, but you cannot rely on this in other libc implementations.
* I would not bother detecting or signalling `fclose` failure. The caller generally has no reasonable response to failure to release a resource like this other than to continue.
* Similarly it's debatable whether handling `printf` failure is worthwhile. If `printf` fails there is generally something more seriously wrong with the platform from which you can't recover anyway. Having said that, I would definitely compile with maximum `-Wformat` paranoia, and if you're going to wrap `printf` make use of `__attribute__((format(printf, …)))` on your wrapper functions.
* I don't know if this is a relevant concern, but wrapping `printf` often also has a non-trivial performance impact. Compilers like GCC know `printf` as a built-in and can do fairly advanced format-string-aware optimizations. But they also generally refuse to inline and specialize varargs functions. Putting these two properties together, wrapping `printf` often thwarts optimizations the compiler would have otherwise done, making a `printf` that used to be almost-free now perhaps costly. But obviously measure first before jumping to any conclusions.
