+++ scripts/README.md
@@ -1,0 +1,5 @@
+# Scripts
+
+This folder contains non-core code.
+
+Useful for experiments, benchmarking, demo applications, etc.
