+++ scripts/python_benchmark/README.md
@@ -1,0 +1,7 @@
+# Python Benchmark
+
+
+
+## Changelog
+
+11-10-2022: Reimplement. Change X Axis to more accurate name
