+++ Makefile
@@ -2,7 +2,7 @@ CODEDIRS=./src/grammar ./src/grammar/rolls
 INCDIRS=./src/grammar
 
 CC=cc
-OPT=-O3 -std=c99 -Wall -Wextra -Werror
+OPT=-O3 -std=c99 -Wall -Wextra -Werror -pedantic -Wcast-align -Wcast-qual -Wctor-dtor-privacy -Wdisabled-optimization -Wformat=2 -Winit-self -Wlogical-op -Wmissing-declarations -Wmissing-include-dirs -Wnoexcept -Wold-style-cast -Woverloaded-virtual -Wredundant-decls -Wshadow -Wsign-conversion -Wsign-promo -Wstrict-null-sentinel -Wstrict-overflow=5 -Wswitch-default -Wundef -Wno-unused
 
 # add flags and the include paths
 CFLAGS=$(foreach D,$(INCDIRS),-I$(D)) $(OPT)
