### 2022-10-11 953cdb2
remove C++ flags
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
