### 2022-10-11 ac4f9a7
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
