### 2022-10-11 cd527cd
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
