### 2022-10-11 3ee6d1b
Update vector_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
