### 2022-10-11 5abfa2b
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
