### 2022-10-11 3359cfc
Update dice_logic.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
