+++ src/grammar/dice_logic.h
@@ -7,6 +7,7 @@
 // Mocking Util
 void reset_mocking();
 void init_mocking(MOCK_METHOD mock_style, int starting_value);
+void mocking_tick();
 
 // Random Functions
 int random_fn(int small, int big );
