### 2022-10-11 cfdda98
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
