### 2022-10-11 cd37e4a
Update dice_logic.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
