### 2022-10-11 04aa331
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
