### 2022-10-11 c6ba861
Update dice_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
