### 2022-10-11 30ab247
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
