### 2022-10-11 b5a68d8
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
