+++ src/grammar/rolls/condition_checking.c
@@ -56,6 +56,11 @@ int check_condition(
             safe_printf("Invalid Conditional\n");
             gnoll_errno = UNDEFINED_BEHAVIOUR;
         }
+        default:{
+            safe_printf("Unknown Conditional\n");
+            gnoll_errno = UNDEFINED_BEHAVIOUR;
+            return 1;
+        }
     }
     safe_printf("Unknown Conditional\n");
     gnoll_errno = UNDEFINED_BEHAVIOUR;
@@ -95,8 +100,13 @@ int check_condition_scalar(
             return 1;
         }
         case INVALID:{
+            safe_printf("Invalid Conditional\n");
+            gnoll_errno = UNDEFINED_BEHAVIOUR;
+        }
+        default:{
             safe_printf("Unknown Conditional\n");
             gnoll_errno = UNDEFINED_BEHAVIOUR;
+            return 1;
         }
     }
     safe_printf("Unknown Conditional\n");
