+++ src/grammar/macro_logic.c
@@ -25,7 +25,7 @@ void register_macro(vec * macro_name, vec *to_store) {
         s->id = key;
         HASH_ADD_INT(macros, id, s);  /* id: name of key field */
     }
-    memcpy(&s->stored_dice_roll, &to_store, sizeof(*to_store));
+    memcpy(&s->stored_dice_roll, &to_store, sizeof(to_store));
 }
 
 struct macro_struct *search_macros(char * skey, vec *to_store) {
