+++ src/grammar/vector_functions.h
@@ -7,8 +7,9 @@
 
 void initialize_vector(vec * vector, DIE_TYPE dt, unsigned int number_of_items);
 
-void pop(int * arr, int len, int value, int * new_arr);
+void pop(int * arr, unsigned int len, int value, int * new_arr);
 
+int contains(int * arr, int len, int value);
 int min(int * arr, int len);
 int max(int * arr, int len);
 
@@ -22,6 +23,7 @@ unsigned int remove_if_present(char ** arr1, int len1,
 
 void collapse_vector(vec * vector, vec * new_vector);
 
+unsigned int keep_logic(vec * vector, vec * new_vector, unsigned int number_to_keep, int keep_high);
 unsigned int keep_highest_values(vec * vector, vec * new_vector, unsigned int number_to_keep);
 unsigned int keep_lowest_values(vec * vector, vec * new_vector, unsigned int number_to_keep);
 unsigned int drop_highest_values(vec * vector, vec * new_vector, unsigned int number_to_keep);
