### 2022-10-11 9779a5a
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
