+++ tests/python/test_macros.py
@@ -27,7 +27,7 @@ def test_D66():
 def test_builtins():
     # Check that builtins are valid calls
     here = os.path.dirname(os.path.abspath(__file__))
-    d = os.path.join(here,"../../builtins/*.dice)
+    d = os.path.join(here,"../../builtins/*.dice")
     for name in glob.glob(d):
        with open(name) as f:
            for macro in f.readlines():
