### 2022-10-11 4300c2b
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
