+++ tests/python/test_macros.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+from glob import glob
 import os
 import pytest
 from util import Mock, roll
