### 2022-10-11 e73b28f
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
