+++ src/grammar/yacc_header.h
@@ -5,9 +5,9 @@
 int initialize();
 // int resolve_dice(dice die);
 
-int sum(int * arr, int len);
-int collapse(int * arr, int len);
+int sum(int * arr, unsigned int len);
+int collapse(int * arr, unsigned int len);
 int roll_numeric_die(int small, int big);
-int roll_symbolic_die(int length_of_symbolic_array);
+int roll_symbolic_die(unsigned int length_of_symbolic_array);
 
 #endif
