### 2022-10-11 4bdd2c1
remove validate_roll, enforced by unsigned ness
