+++ src/grammar/dice_logic.c
@@ -145,14 +145,3 @@ int * do_roll(roll_params rp){
         rp.start_value
     );
 }
-
-void validate_roll(int number_of_dice, int die_side)
-{
-    if (die_side < 0){
-        printf("Cannot roll a dice with a negative amount of sides\n");
-        gnoll_errno = INTERNAL_ASSERT;
-    }if (number_of_dice < 0){
-        printf("Cannot roll a negative number of dice\n");
-        gnoll_errno = INTERNAL_ASSERT;
-    }
-}
