+++ src/grammar/macro_logic.h
@@ -10,7 +10,7 @@
 
 struct macro_struct {
     int id;                    /* key */
-    // char name[MAX_SYMBOL_TEXT_LENGTH];
+    // char name[MAX_SYMBOL_LENGTH];
     vec stored_dice_roll;
     UT_hash_handle hh;         /* makes this structure hashable */
 };
