+++ README.md
@@ -1,89 +1,39 @@
-# PetMe
+## Scripts
 
-<h3>Join official <a href="https://discord.gg/X3HZAPzykU">Discord Channel</a> for discussion</h3>
+Here is the list of scripts available:
 
-## Inspiration💡
+- [Sitemap](#sitemap)
+- [Roadmaps Meta](#roadmaps-meta)
+- [Content Skeleton](#content-skeleton)
 
-Each year, it's estimated that more than one million adoptable dogs and cats are euthanized in the United States, simply because too many pets come into shelters and too few people consider adoption when looking for a pet.
+## Sitemap
 
-**The number of euthanized animals could be reduced dramatically if more people adopt pets instead of buying them. When you adopt dogs and cats, you save a loving animal by making them part of your family and open up shelter space for another animal who might desperately need it.**
+Generates the sitemap with all the pages and guides
 
-We are following the policy of saving lives, saving animals and we built **PetMe** in this hackathon.
+```shell
+npm run meta:sitemap
+```
 
-## What it does⚙️
+## Roadmaps Meta
 
-### Test here [PetMe](https://akshitagupta15june.github.io/PetMe/)
+Generates the `content/roadmaps.json` file by combining the `content/raodmaps/**/meta.json` content in each roadmap
 
-**Our project is PetMe, a platform that allows animals to be adopted, donated to pet lovers, and provides emergency medical care to stray animals in need.**
+```shell
+npm run meta:roadmaps
+```
 
-🤖 The first feature is **Animal Adoption**, which allows users to choose an animal they want to adopt and submit their details so that the owner can reach out to them regarding adoption details. 
+## Content Skeleton
 
-🤖 The next feature is **Pet Donation**. In these hectic times, pet owners may not be able to provide proper care to their pets because of their business. In such instances, they donate the pet to another person who can adopt the animal.
+This command is used to create the content folders and files for the interactivity of the roadmap. You can use the below command to generate the roadmap skeletons inside a roadmap directory:
 
-🤖 All the data will be stored in a database and as soon as a person add the details of a pet animal they want to donate so their details will be shown in the adoption menu.
+```shell
+npm run roadmap-content [frontend|backend|devops|...]
+```
 
-🤖  Our platform also provides **SOS reports**, as the population of stray animals is increasing, and stray animals cannot stand harsh weather conditions, and sometimes they are injured by accidents. Therefore, if a person notices any stray animal in harsh conditions, they can fill out the details and a representative will reach the location to assist the animal with medical aid. 
+For the content skeleton to be generated, we should have proper grouping, and the group names in the project files. You can follow the steps listed below in order to add the meta information to the roadmap. 
 
-🤖  Users can also **contact us** through the contact us form and one of our team members will contact them shortly.
-
-
-## How we built it ⚙️
-
-We have built this website using HTML, CSS, JavaScript and connected backend database with postman API's.
-
-## Accomplishments that we're proud of 😇
-
-We are proud that we built a website in which pet lovers can easily adopt stray animals through our website, which not only provides medical assistance to stray animals, but allows them to be adopted by their comrades in the community.
-
-## What's next for PetMe 📲
-
-In the future, we plan to scale this application more to reach as many people as possible and facilitate adoption and donation of animals.
-
[…]
