### 2022-10-11 383158b
turnoff errors for files out of my control
