### 2022-10-11 a4b8991
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
