### 2022-10-11 174304b
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
