### 2022-10-11 e7b4840
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
