### 2022-10-11 dc5520c
Update macro_logic.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
