### 2022-10-11 015c39e
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
