+++ Makefile
@@ -3,10 +3,12 @@ INCDIRS=./src/grammar
 
 CC=cc
 OPT=-O3 -std=c99 -Wall -Wextra -Werror -pedantic -Wcast-align \
-	-Wcast-qual -Wdisabled-optimization -Wformat=2 -Winit-self \
+	-Wcast-qual -Wdisabled-optimization -Winit-self \
 	-Wmissing-declarations -Wmissing-include-dirs \
 	-Wredundant-decls -Wshadow -Wsign-conversion \
 	-Wundef -Wno-unused
+# To be enabled after printf investigation
+# -Wformat=2
 
 # YACC/LEX fails for the following, so disabled:
 # -Wswitch-default  -Wstrict-overflow=5
