### 2022-10-11 db5b58f
Update safe_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
