+++ src/grammar/safe_functions.c
@@ -30,7 +30,8 @@ void * safe_calloc(size_t nitems, size_t size){
     }
     void * calloc_result;
     calloc_result = calloc(nitems, size);
-    if(!calloc_result && (nitems*size)){
+    unsigned int total_sz = nitems*size;
+    if(!calloc_result && total_sz){
         gnoll_errno = BAD_ALLOC;
     }
     return calloc_result;
