### 2022-10-11 5284544
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
