### 2022-10-11 b0d0001
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
