### 2022-10-12 57c1638
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
