### 2022-10-12 4decba8
Update test_macros.py
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
