+++ tests/python/test_legacy.py
@@ -7,5 +7,6 @@ from util import Mock, roll
     ("2d2d2d==1", 3, Mock.RETURN_INCREMENTING, 1),
 ])
 def test_rerolling_multidice(r, out, mock, mock_const):
-    result = roll(r, mock_mode=mock, mock_const=mock_const)
-    assert result == out
+    with pytest.raises(Exception):
+        result = roll(r, mock_mode=mock, mock_const=mock_const)
+        assert result == out
