+++ tests/python/util.py
@@ -34,7 +34,8 @@ def get_roll():
 
 def make_all():
     cmd = "make all -s -C " + MK_DIR
-    parser = subprocess.Popen(cmd, shell=True)
+    cmd = cmd.split(' ')
+    parser = subprocess.Popen(cmd, shell=False)
     parser.communicate()
     if parser.returncode:
         raise ValueError
@@ -55,5 +56,5 @@ def roll(s, mock_mode=Mock.NO_MOCK, mock_const=3):
 
     if exit_code:
         raise ValueError
-        
+
     return result
