+++ .history/doc/gh-pages/_config_20221012223347.yml
@@ -14,9 +14,11 @@ callouts:
   warning:
     title: ⚠️ Warning
     color: red
+  
   explanation:
     title: 🤔 Design Decision
     color: blue
+
   story:
     title: ⚔️ Grindon's Tale
     color: yellow
