+++ .history/.github/workflows/pip_test_20221012223500.yml
@@ -1,4 +1,4 @@
-name: "Test: C/C++/Python"
+name: "Test: Packaging: PyPi"
 
 on:
   push:
@@ -7,21 +7,20 @@ on:
     branches: [ main ]
 
 jobs:
-  build:
+  verify:
     runs-on: ubuntu-latest
 
     steps:
+    
     - uses: actions/checkout@v3
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
         python-version: '3.x'
     - uses: actions/checkout@v2
-    - name: Install dependencies
+    - name: Test PyPi 
       run: |
-        python -m pip install --upgrade pip
-        python -m pip install -r reqs/requirements.txt
-        pip install build
-        make all
-    - name: make test
-      run: make test
+        python -m pip install gnoll
+        ls /opt/hostedtoolcache/Python/3.10.7/x64/lib/python3.10/site-packages/gnoll/
+        python -c "from gnoll.parser import roll; roll('d34')"
+    
