+++ .history/doc/gh-pages/_config_20221012222315.yml
@@ -14,12 +14,15 @@ callouts:
   warning:
     title: ⚠️ Warning
     color: red
+  
   explanation:
     title: 🤔 Design Decision
     color: blue
+  
   story:
     title: ⚔️ Grindon's Tale
     color: yellow
+  
   gnoll:
     title: 🐺 GNOLL
     color: grey-lt
