+++ .history/.github/workflows/pip_test_20221012223747.yml
@@ -1,25 +1,0 @@
-name: "Test: Packaging: PyPi"
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  verify:
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python
-      uses: actions/setup-python@v3
-      with:
-        python-version: '3.x'
-    - uses: actions/checkout@v2
-    - name: Test PyPi
-      run: |
-        python -m pip install gnoll
-        ls /opt/hostedtoolcache/Python/3.10.7/x64/lib/python3.10/site-packages/gnoll/
-        python -c "from gnoll.parser import roll; roll('d34')"
-    
