+++ .history/.github/workflows/c-cpp_20221012223421.yml
@@ -1,28 +1,0 @@
-name: "Test: C/C++/Python"
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python
-      uses: actions/setup-python@v3
-      with:
-        python-version: '3.x'
-        
-    - uses: actions/checkout@v2
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install -r reqs/requirements.txt
-        pip install build
-        make all
-    - name: make test
-      run: make test
