+++ .history/doc/gh-pages/_config_20221012223348.yml
@@ -1,26 +1,0 @@
-# See: https://just-the-docs.github.io/just-the-docs/
-remote_theme: just-the-docs/just-the-docs
-plugins:
-  - jekyll-seo-tag
-  - jekyll-remote-theme
-
-# Search
-search_enabled: true
-search.heading_level: 3
-search.button: true
-
-# Callout blocks
-callouts:
-  warning:
-    title: ⚠️ Warning
-    color: red
-  
-  explanation:
-    title: 🤔 Design Decision
-    color: blue
-  story:
-    title: ⚔️ Grindon's Tale
-    color: yellow
-  gnoll:
-    title: 🐺 GNOLL
-    color: grey-lt

[…]
