### 2022-10-12 46979bb
Apply suggestions from code review.
Co-authored-by: Ian Hunter <ian.hunter@intel.com>
Signed-off-by: Olivier <kineolyan@proton.me>
