+++ doc/gh-pages/notation/custom_macros.md
@@ -7,7 +7,8 @@ nav_order: 10
 # Custom Macros
 
 {: .story }
->**GM**: You wake up from your dream, finding an old creature resembling to a witch. Seeing your bleeding head, it offers you to choose a potion.
+>**GM**: You wake up from your dream, finding an old creature resembling a witch. Seeing your bleeding head, it offers you to choose a potion.
+
 >
 >**Grindon**: Can I pick the healing potion?
 >
@@ -21,7 +22,8 @@ nav_order: 10
 >
 >*GNOLL*: [something else]
 >
->**Grindon The Brave**: I am starting to fell weird...
+>**Grindon The Brave**: I am starting to feel weird...
+
 
 ## Usage
 
