+++ doc/gh-pages/notation/builtin_macros.md
@@ -8,7 +8,7 @@ nav_order: 9
 # Built-In Macros
 
 {: .story }
->**GM**: You fail to secure yourself on the wall and slip towards the spikes. While falling, you hit your head against the wall. 
+>**GM**: You fail to secure yourself on the wall and slip towards the spikes. While falling, you hit your head against the wall.
 >
 >**Grindon**: Gosh, that hurts.
 >
@@ -39,7 +39,7 @@ Here are a sample of the dice that can be used via Macro Accessors
 | Directions | NORTH, SOUTH, WEST, EAST | #DIRECTION_D4 |
 | Card Suits| CLUBS, DIAMONDS, HEARTS, SPADES | #CARD_SUIT |
 | Zodiac | Aries through Pisces | #ZODIAC |
-| Planets | Sun, Moon, Mercury through Pluto, Ascending Node, Descending Node | #PLANETS |
+| Planets | Sun, Moon, all the planets from Mercury all the way to Pluto, Ascending Node, Descending Node | #PLANETS |
 
 More can be found in the `builtins/` folder [here](https://github.com/ianfhunter/GNOLL/tree/main/builtins)
 
