### 2022-10-12 3fa9500
Remote the extra line I used to write the stories.
