### 2022-10-12 b226f5f
Update pip_test.yml
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
