### 2022-10-12 7ab34f5
Simplified test fn
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
