### 2022-10-12 be20b10
Update safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
