+++ src/grammar/safe_functions.h
@@ -20,12 +20,8 @@ typedef enum {
 void * safe_malloc(size_t size);
 void * safe_calloc(size_t nitems, size_t size);
 FILE * safe_fopen(const char *filename, const char *mode);
-int safe_fclose(FILE *stream);
 char * safe_strdup( const char *str1 );
 long int safe_strtol (const char* str, char** endptr, int base);
-
-void safe_fprintf(FILE *stream, const char *format, ...);
-void safe_printf(const char *fmt, ...);
 
 void end_gnoll(int err_code);
 
