### 2022-10-12 5402f00
Update safe_functions.h
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
