### 2022-10-12 fc7f641
Update vector_functions.c
Signed-off-by: Ian Hunter <ianfhunter@gmail.com>
