### 2022-10-16 ianfhunter/GNOLL#292 @Smattr
Re-rolling in quick succession generates the same repeated result
**Describe the bug**

Running `dice` rapidly in a loop leads to long strings of the same dice roll result.

**To Reproduce**

If you try something like the following quickly, you will notice you get the same dice roll repeating:

```
$ dice 1d20
13;
$ dice 1d20
13;
$ dice 1d20
13;
$ dice 1d20
13;
$ dice 1d20
4;
$ dice 1d20
4;
$ dice 1d20
4;
$ dice 1d20
4;
$ dice 1d20
4;
$ dice 1d20
9;
```

**Expected behavior**

Dice rolling, even in rapid succession, should generate statistically random results.

**Environment (please complete the following information):**

I don't think this is OS-dependent, but:

```
$ uname -rms
Linux 5.15.0-50-generic x86_64
$ lsb_release -a
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 22.04.1 LTS
Release:        22.04
Codename:       jammy
```

**Additional context**

I believe what's happening here is that the random number generator is seeded with the current time (`time(0)`). So any two executions that occur within the same second will get identical RNG state. This means the problem is not just limited to a one-off roll (asking for several rolls will likely generate the same repeated _sequence_ of rolls) or to serial execution (two `dice` processes running in parallel will return the same result).

Using the system clock to seed an RNG is a fairly pervasive practice, but also cryptographically unsound. There is a distinction in security literature between Pseudo Random Number Generators (PRNGs) and a sub-category of Crytographically Secure Pseudo Random Number Generators (CSPRNGs), with the latter needing more care when it comes to both seeding and the algorithm of choice. (I'm probably explaining things you already know, but I just wanted to be clear for anyone else spectating this issue).

Dice rollers like GNOLL fall in this interesting space where I'm not really sure if they need a CSPRNG or not – if a malicious party can control and/or predict your dice roll, is that a problem? This seems very application-specific. That is, GNOLL _itself_ is basically an RNG, and some applications using it will answer “yes” to the previous question and others will answer “no.” However, CSPRNGs these days are easily available and fast. So it’s probably a good idea to switch to using one. The applications who said “yes” will be happy and the ones who said “no” won’t care. I don’t know what the options are on Windows, but on Linux you have `getrandom` on modern kernels.

It might also be a good idea to introduce a test case that checks the outputs are statistically random. Though this is a little harder to construct.

Please let me know if you have questions or if I’ve misunderstood something.
