### 2022-10-16 ianfhunter/GNOLL#293 @Smattr
a negative number of dice can be rolled
**Describe the bug**
In the `xdy` notation of the input, it's possible to give a negative number for `x`. This does not error out, but returns a (negative) result. It's unclear how to interpret this.

**To Reproduce**
```
$ dice -9d20
-105;
```

**Expected behavior**

An error? Or is a negative number notation for something else specific? I'm not particularly knowledgeable about dice rolling notation, so it's possible `-9d20` actually means something perfectly sensible to the experts.

**Environment (please complete the following information):**

```
$ uname -rms
Linux 5.15.0-50-generic x86_64
$ lsb_release -a
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 22.04.1 LTS
Release:        22.04
Codename:       jammy
```

**Additional context**

N/A
