### 2022-10-16 ianfhunter/GNOLL#293 @Smattr
Oh, apparently I previously came across this in https://github.com/ianfhunter/GNOLL/issues/249. I guess my experience just now suggests https://github.com/ianfhunter/GNOLL/pull/253 didn't quite fix it though.
