### 2022-10-16 #4816 @Smattr
Sorry for the delay. I've been making my way through the checklist now, filing issues as I go:
* https://github.com/ianfhunter/GNOLL/issues/249
* https://github.com/ianfhunter/GNOLL/issues/251
* https://github.com/ianfhunter/GNOLL/issues/252
* https://github.com/ianfhunter/GNOLL/issues/292
* https://github.com/ianfhunter/GNOLL/issues/293

Ian's been addressing them faster than I file them, so things going well!

The remaining item on my checklist is the performance claims. This paper benchmarks against two proprietary tools I don't have access to. What are the expectations for reviewing in this situation?
