### 2022-10-17 #4816 @ianfhunter
> The remaining item on my checklist is the performance claims. This paper benchmarks against two proprietary tools I don't have access to. What are the expectations for reviewing in this situation?

I noticed this on the checklist, so I have been creating a github action to make it easier to replicate
https://github.com/ianfhunter/GNOLL/actions/runs/3259062234

The python comparison is ready.
The other performance claim is for the following libraries:
- https://github.com/EBailey67/DiceParser
- http://hjemmesider.diku.dk/~torbenm/Troll/Troll.zip

However, when measuring Troll, I edited it heavily locally to remove print statements and extra statistical logic (to give a fairer comparison). I am unsure of the licensing obligations of the library were I to upload a fork/patch. I have the unmodified version on a PR.
