### 2022-10-17 ianfhunter/GNOLL#293 @ianfhunter
I believe what is happening here is `-(9d20)` rather than `(-9)d20`. It is a began of the overall result, not a rolling of a negative amount of dice.

I don't know of any circumstances that you would like to roll a negative value, but it's enabled as part of the general arithmetic capabilities (e.g. 2d10-d6). So I believe this is working as intended
