### 2022-10-17 ianfhunter/GNOLL#293 @Smattr
I don't quite understand your use of the word "began" there, but overall I get that you're saying it's not a bug and working as designed. Apologies for the noise!
