### 2022-10-17 ianfhunter/GNOLL#292 @ianfhunter
Thanks for this, these were both great catches. I hadn't really thought about the security of the random function before, but I could certainly think of some use cases that would require it. Mostly for gambling-based applications.

I had a suspicion that rand() was not great for statistical consistency, but never really thought to investigate it. PCG seems to give us an improvement on this (though i have not individually verified yet). I added Arc4Random as an optional switch, because all the secure algorithms seem to come at a performance cost. (Not that GNOLL is hurting on performance, but I imagine the average user would prefer the statistical consistency.)
