### 2022-10-17 ianfhunter/GNOLL#292 @ianfhunter
Out of interest I ran a normality test and these other two methods are consistently performing 10x better in terms of statistical consistency:

(>2 dice rolled yield approximate normal distribution, increasingly so as the numbers of dice increase) [reference](https://math.stackexchange.com/questions/406192/probability-distribution-of-rolling-multiple-dice)

![image](https://user-images.githubusercontent.com/1540737/196222185-8e320506-b7bc-47c4-9838-7ab9c94ece21.png)
![image](https://user-images.githubusercontent.com/1540737/196222451-4c3414e5-911f-44e5-ba5f-aae245bbc051.png)
![image](https://user-images.githubusercontent.com/1540737/196222625-8b65f2be-75d5-4710-8721-95dacb27946c.png)

After a few runs on each, it seems the PCG summary is about right, that PCG>Arc4Random>rand()
