### 2022-10-17 ianfhunter/GNOLL#293 @ianfhunter
Autocorrect! I meant negation
