### 2022-10-23 #4816 @Smattr
I get slightly different results, but they look within expected bounds. Given these tests are measuring wall clock time rather than cycles, I assume some variance is expected. It looks like d20 has been added in the meantime and the range of the C benchmark has been reduced.

![py](https://user-images.githubusercontent.com/203893/197403822-55fda742-4a97-4382-8f75-d4aed502f3f7.PNG)
![C++](https://user-images.githubusercontent.com/203893/197403827-af40478e-655e-4364-95a8-2f6e8054f839.PNG)
