### 2022-11-22 #4816 @ianfhunter
I had another look at the spreadsheet. Here are a few that may be applicable if you have not invited them already (in descending order)

- [petar](https://bitbucket.org/petar/) (Knowledge of Parsers)
- zbeekman (Build systems & package Management)
- rlskoeser (Data/Text Parsing, though may be in different context, no c/c++)
- corybrunson (Packaging, though only R Lang)
