### 2022-12-17 #4816 @Smattr
> ianfhunter I just sent invites to the two first recommendations on your list since I think we do need someone that can assess the C/C++ parts of your project. I have also sent out emails to two additional people that I think have the right skills and domain knowledge. Thanks again for your patience.
>
> danielskatz it's been pretty hard to find C/C++ reviewers for this domain, if you know of anyone that is available to volunteer we'd really appreciate it.

In case it wasn't clear, the parts I assessed in my review covered the C/C++ components too.
