### 2022-12-28 #4816 @steven-murray
@ianfhunter this is a really cool library!

I've made a few issues, mostly concerning documentation, but otherwise things look great. The main part of the documentation (Grindon's Tale) is very clear and even interesting :-).

I will note that the PDF of the actual paper linked above does have a reference to "Anonymous" that should be fixed,
