### 2023-01-04 #4816 @galessiorob
@ianfhunter great work on those final changes! And thank you, @steven-murray, for that last check.

I put together some edit suggestions on the actual paper [here](https://github.com/ianfhunter/GNOLL/commit/9a33db56db011552aa8bcf48bdde46634f1aa625), please include them as you see fit. Also, it looks like the "citations" section is still empty? You can add this section and/or a references one after the figures section. Let me know what you think.

Almost there!
