### 2023-01-04 #4816 @ianfhunter
> I put together some edit suggestions on the actual paper [here](https://github.com/ianfhunter/GNOLL/commit/9a33db56db011552aa8bcf48bdde46634f1aa625), please include them as you see fit. Also, it looks like the "citations" section is still empty? You can add this section and/or a references one after the figures section. Let me know what you think.

Thank you, I will review those proposed changes as soon as I can (currently abroad).

I used the example paper as a template when creating the paper, but I'm not really sure what the difference between the citations section and references should be? I believe there is a populated references section in the current document (at the end after figures), but if only one of the two is required, I can delete the blank "citations" section
