### 2023-01-16 #4816 @galessiorob
Apologies for missing your email @ianfhunter, I think we're ready for the final steps! 🚀

The next thing we need from you is to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

@steven-murray and @Smattr thank you so, so much for your thorough and thoughtful reviews 🙏
