### 2023-01-16 #4816 @ianfhunter
**Zenodo:** https://zenodo.org/record/7542433
**DOI:** 10.5281/zenodo.7542433
**Latest GitHub Release (today):** https://github.com/ianfhunter/GNOLL/releases/tag/v4.3.2

I hope this is what you're looking for.

@steven-murray , @Smattr - I also want to voice my appreciation for your reviews and sharing your knowledge. I've certainly learned a lot and the project is much better for it!

And thank you too @galessiorob for orchestrating it all!
