### 2023-01-17 #4816 @ianfhunter
> @ianfhunter - I'm the track editor who will finish the processing of your submission. I've suggested some small edits in [ianfhunter/GNOLL#434](https://github.com/ianfhunter/GNOLL/pull/434). Please merge this, or let me know what you disagree with.
>
> Also, the sentence that ends with "many software developers have created their own dice notation parsers (for a simple subset is not too difficult to develop)" is not clear to me as written. Could this be changed to "many software developers have created their own dice notation parsers (because a parser for a simple subset is not too difficult to develop)"? Or am I misreading this?

Thank you, I've applied those suggestions, including the clarification. You're absolutely correct, it reads much easier this way
